API

Wavelets.Threshold.VisuShrink

WaveletsExt.Denoising.RelErrorShrink

WaveletsExt.Denoising.RelErrorShrink

WaveletsExt.Denoising.SureShrink

WaveletsExt.Denoising.SureShrink

Wavelets.Threshold.bestbasistree

Wavelets.Threshold.denoise

Wavelets.Threshold.noisest

WaveletsExt.ACWT.acwpt

WaveletsExt.ACWT.acwt

WaveletsExt.ACWT.acwt_step

WaveletsExt.ACWT.autocorr

WaveletsExt.ACWT.hacwt

WaveletsExt.ACWT.iacwt

WaveletsExt.ACWT.make_acqmfpair

WaveletsExt.ACWT.make_acreverseqmfpair

WaveletsExt.ACWT.pfilter

WaveletsExt.ACWT.qfilter

WaveletsExt.ACWT.vacwt

WaveletsExt.BestBasis.bestbasis_treeselection

WaveletsExt.BestBasis.bestbasiscoef

WaveletsExt.BestBasis.tree_costs

WaveletsExt.Denoising.denoiseall

WaveletsExt.Denoising.findelbow

WaveletsExt.Denoising.orth2relerror

WaveletsExt.Denoising.relerrorthreshold

WaveletsExt.Denoising.surethreshold

WaveletsExt.LDB.discriminant_measure

WaveletsExt.LDB.discriminant_power

WaveletsExt.LDB.energy_map

WaveletsExt.LDB.ldb

WaveletsExt.SIWPD.makesiwpdtree

WaveletsExt.SIWPD.siwpd

WaveletsExt.SWT.isdwt

WaveletsExt.SWT.isdwt_step

WaveletsExt.SWT.isdwt_step!

WaveletsExt.SWT.iswpt

WaveletsExt.SWT.sdwt

WaveletsExt.SWT.sdwt_step

WaveletsExt.SWT.sdwt_step!

WaveletsExt.SWT.swpd

WaveletsExt.SWT.swpt

WaveletsExt.Utils.coarsestscalingrange

WaveletsExt.Utils.finestdetailrange

WaveletsExt.Utils.generatesignals

WaveletsExt.Utils.getleaf

WaveletsExt.Utils.left

WaveletsExt.Utils.nodelength

WaveletsExt.Utils.psnr

WaveletsExt.Utils.relativenorm

WaveletsExt.Utils.right

WaveletsExt.Utils.snr

WaveletsExt.Utils.ssim

WaveletsExt.Visualizations.plot_tfbdry

WaveletsExt.Visualizations.treenodes_matrix

WaveletsExt.Visualizations.wiggle

WaveletsExt.Visualizations.wiggle!

WaveletsExt.WPD.wpd

WaveletsExt.WPD.wpd!

WaveletsExt.ACWT.acwpt — Function

acwpt(x, wt[, L=maxtransformlevels(x)])

Performs a discrete autocorrelation wavelet packet transform for a given signal x. The wavelet type wt determines the transform type. Refer to Wavelet.jl for a list of available methods.

Examples

acwpt(x, wavelet(WT.db4))

acwpt(x, wavelet(WT.db4), 4)

See also: acwt, acwpt_step, iacwpt

WaveletsExt.ACWT.acwt — Function

acwt(x, wt[, L=maxtransformlevels(x)])

acwt(x, wt[, Lrow=maxtransformlevels(x[1,:]), Lcol=maxtransformlevels(x[:,1])])

Performs a discrete autocorrelation wavelet transform for a given signal x. The signal can be 1D or 2D. The wavelet type wt determines the transform type. Refer to Wavelet.jl for a list of available methods.

Examples

acwt(x, wavelet(WT.db4))

acwt(x, wavelet(WT.db4), 4) # level 4 decomposition

See also: acwt_step, iacwt

WaveletsExt.ACWT.acwt_step — Function

acwt_step(v, j, h, g)

Performs one level of the autocorrelation discrete wavelet transform (ACWT) on the vector v, which is the j-th level scaling coefficients (Note the 0th level scaling coefficients is the raw signal). The vectors h and g are the detail and scaling filters.

Returns a tuple (v, w) of the scaling and detail coefficients at level j+1.

See also: acwt, iacwt

WaveletsExt.ACWT.autocorr — Method

autocorr(f::OrthoFilter)

Generates the autocorrelation filter for a given wavelet filter.

WaveletsExt.ACWT.hacwt — Function

hacwt(x, wt[, L=maxtransformlevels(x)])

Computes the column-wise discrete autocorrelation transform coeficients for 2D signals.

WaveletsExt.ACWT.iacwt — Function

iacwt(xw::AbstractArray{<:Number,2})

iacwt(xw::AbstractArray{<:Number,4})

Performs the inverse autocorrelation discrete wavelet transform. Can be used for both the 1D and 2D case.

See also: iacwt!, acwt

WaveletsExt.ACWT.make_acqmfpair — Method

acmakeqmfpair(f::OrthoFilter)

Generates the autocorrelation quadratic mirror filters.

WaveletsExt.ACWT.make_acreverseqmfpair — Method

makereverseqmfpair(f::OrthoFilter)

Generates the reverse autocorrelation quadratic mirror filters.

WaveletsExt.ACWT.pfilter — Method

pfilter(f::OrthoFilter)

Generates the high-pass autocorrelation filter

WaveletsExt.ACWT.qfilter — Method

qfilter(f::OrthoFilter)

Generates the low-pass autocorrelation filter

WaveletsExt.ACWT.vacwt — Function

vacwt(x, wt[, L=maxtransformlevels(x)])

Computes the row-wise discrete autocorrelation transform coeficients for 2D signals.

Wavelets.Threshold.bestbasistree — Method

bestbasistree(X[, method])

Extension to the best basis tree function from Wavelets.jl. Given a set of decomposed signals, returns different types of best basis trees based on the methods specified. Available methods are the joint best basis (JBB()), least statistically dependent basis (LSDB()), and individual regular best basis (BB()).

WaveletsExt.BestBasis.bestbasis_treeselection — Method

bestbasis_treeselection(costs, n)

Computes the best tree based on the given cost vector.

WaveletsExt.BestBasis.bestbasiscoef — Method

bestbasiscoef(X, tree)
bestbasiscoef(X, wt, tree)

Returns the expansion coefficients based on the given tree(s) and wavelet packet decomposition (WPD) expansion coefficients. If the WPD expansion coefficients were not provided, the expansion coefficients can be obtained by providing the signals and wavelet.

WaveletsExt.BestBasis.tree_costs — Method

tree_costs(X, method)

Returns the cost of each node in a binary tree in order to find the best basis.

Wavelets.Threshold.denoise — Method

denoise(x, inputtype, wt[; L=maxtransformlevels(size(x,1)),
    tree=maketree(size(x,1), L, :dwt), dnt=VisuShrink(size(x,1)),
    estnoise=noisest, smooth=:regular])

Extension of the denoise function from Wavelets.jl. Denoise a signal of input type inputtype.

Arguments:

x::AbstractArray{<:Number}: input signals/coefficients.

inputtype::Symbol: input type of x. Current accepted types of inputs are

:sig: original signals; x should be a 2-D array with each column representing a signal.

:dwt: dwt-transformed signal coefficients; x should be a 1-D array with each column representing the coefficients of a signal.

:wpt: wpt-transformed signal coefficients; x should be a 1-D array with each column representing the coefficients of a signal.

:sdwt: sdwt-transformed signal coefficients; x should be a 2-D array with each column representing the coefficients of a node.

:swpd: swpd-transformed signal coefficients; x should be a 2-D array with each column representing the coefficients of a node.

:acwt: acwt-transformed signal coefficients from AutocorrelationShell.jl; x should be a 2-D array with each column representing the coefficients of a node.

:acwpt: acwpt-transformed signal coefficients from AutocorrelationShell.jl; x should be a 2-D array with each column representing the coefficients of a node.

wt::Union{DiscreteWavelet, Nothing}: the discrete wavelet to be used for decomposition (for input type :sig) and reconstruction. nothing can be supplied if no reconstruction is necessary.

L::Integer: the number of decomposition levels. Necessary for input types :sig, :dwt, and :sdwt. Default value is set to be maxtransformlevels(size(x,1)).

tree::BitVector: the decomposition tree of the signals. Necessary for input types :wpt and :swpd. Default value is set to be maketree(size(x,1), L, :dwt).

dnt::DNFT: denoise type. Default type is set to be VisuShrink(size(x,1)).

estnoise::Union{Function, Vector{<:Number}}: noise estimation. Input can be provided as a function to estimate noise in signal, or a vector of estimated noise. Default is set to be the noisest function.

smooth::Symbol: the smoothing method used. :regular smoothing thresholds all given coefficients, whereas :undersmooth smoothing does not threshold the lowest frequency subspace node of the wavelet transform. Default is set to be :regular.

See also: denoiseall

Wavelets.Threshold.noisest — Method

noisest(x, redundant[, tree=nothing])

Extension to the noisest function from Wavelets.jl. Estimates the noise of a signal from its coefficients.

Examples

x = randn(128)
wt = wavelet(WT.haar)

# noise estimate for dwt transformation
y = dwt(x, wt)
noise = noisest(y, false)

# noise estimate for wpt transformation
tree = maketree(x, :full)
y = wpt(x, wt, tree)
noise = noisest(y, false, tree)

# noise estimate for sdwt transformation
y = sdwt(x, wt)
noise = noisest(y, true)

# noise estimate for swpd transformation
y = swpd(x, wt)
noise = noisest(y, true, tree)

See also: relerrorthreshold

WaveletsExt.Denoising.denoiseall — Method

denoiseall(x, inputtype, wt[; L=maxtransformlevels(size(x,1)),
    tree=maketree(size(x,1), L, :dwt), dnt=VisuShrink(size(x,1)),
    estnoise=noisest, bestTH=nothing, smooth=:regular])

Denoise multiple signals of input type inputtype.

Arguments:

x::AbstractArray{<:Number}: input signals/coefficients.

inputtype::Symbol: input type of x. Current accepted types of inputs are

:sig: original signals; x should be a 2-D array with each column representing a signal.

:dwt: dwt-transformed signal coefficients; x should be a 2-D array with each column representing the coefficients of a signal.

:wpt: wpt-transformed signal coefficients; x should be a 2-D array with each column representing the coefficients of a signal.

:sdwt: sdwt-transformed signal coefficients; x should be a 3-D array with each 2-D slice representing the coefficients of a signal.

:swpd: swpd-transformed signal coefficients; x should be a 3-D array with each 2-D slice representing the coefficients of a signal.

:acwt: acwt-transformed signal coefficients from AutocorrelationShell.jl; x should be a 3-D array with each 2-D slice representing the coefficients of a signal.

:acwpt: acwpt-transformed signal coefficients from AutocorrelationShell.jl; x should be a 3-D array with each 2-D slice representing the coefficients of a signal.

wt::Union{DiscreteWavelet, Nothing}: the discrete wavelet to be used for decomposition (for input type :sig) and reconstruction. nothing can be supplied if no reconstruction is necessary.

L::Integer: the number of decomposition levels. Necessary for input types :sig, :dwt, and :sdwt. Default value is set to be maxtransformlevels(size(x,1)).

tree::BitVector: the decomposition tree of the signals. Necessary for input types :wpt and :swpd. Default value is set to be maketree(size(x,1), L, :dwt).

dnt::DNFT: denoise type. Default type is set to be VisuShrink(size(x,1)).

estnoise::Union{Function, Vector{<:Number}}: noise estimation. Input can be provided as a function to estimate noise in signal, or a vector of estimated noise. Default is set to be the noisest function.

bestTH::Union{Function, Nothing}: method to determine the best threshold value for a group of signals. If nothing is given, then each signal will be denoised by its respective best threshold value determined from the parameters dnt and estnoise; otherwise some function can be passed to determine the best threshold value from a vector of threshold values, eg: mean and median. Default is set to be nothing.

smooth::Symbol: the smoothing method used. :regular smoothing thresholds all given coefficients, whereas :undersmooth smoothing does not threshold the lowest frequency subspace node of the wavelet transform. Default is set to be :regular.

See alse: denoise, noisest

WaveletsExt.Denoising.findelbow — Method

findelbow(x, y)

Given the x and y coordinates of a curve, return the elbow.

WaveletsExt.Denoising.orth2relerror — Method

orth2relerror(orth)

Given a vector 'orth' of orthonormal expansion coefficients, return a vector of relative approximation errors when retaining the 1,2,...,N largest coefficients in magnitude.

WaveletsExt.Denoising.relerrorthreshold — Method

relerrorthreshold(coef, redundant[, tree, elbows=2; makeplot=false])

Takes in a set of expansion coefficients, 'plot' the threshold vs relative error curve and select the best threshold value based on the elbow method.

Examples

x = randn(128)
wt = wavelet(WT.haar)

# noise estimate for dwt transformation
y = dwt(x, wt)
noise = relerrorthreshold(y, false)

# noise estimate for wpt transformation
tree = maketree(x, :full)
y = wpt(x, wt, tree)
noise = relerrorthreshold(y, false, tree)

# noise estimate for sdwt transformation
y = sdwt(x, wt)
noise = relerrorthreshold(y, true)

# noise estimate for swpd transformation
y = swpd(x, wt)
noise = relerrorthreshold(y, true, tree)

See also: noisest, RelErrorShrink

WaveletsExt.Denoising.surethreshold — Method

surethreshold(coef, redundant[, tree=nothing])

Determination of the t value used for SureShrink.

See also: SureShrink

Wavelets.Threshold.VisuShrink — Method

VisuShrink(n, th)

Extension to the VisuShrink struct constructor from Wavelets.jl.

WaveletsExt.Denoising.RelErrorShrink — Type

Relative Error Shrink

WaveletsExt.Denoising.RelErrorShrink — Type

RelErrorShrink([th=HardTH()])

Struct constructor for Relative Error Shrink.

WaveletsExt.Denoising.SureShrink — Type

Stein's Unbiased Risk Estimate (SURE) Shrink

WaveletsExt.Denoising.SureShrink — Type

SureShrink(x[, tree=nothing, th=SteinTH()])

Struct constructor for SureShrink based on the signal coefficients x.

WaveletsExt.LDB.discriminant_measure — Method

discriminant_measure(Γ, dm)

Returns the discriminant measure calculated from the energy maps.

WaveletsExt.LDB.discriminant_power — Method

discriminant_power(coefs, tree, dp)

Returns the discriminant power of each leaf from the local discriminant basis (LDB) tree.

WaveletsExt.LDB.energy_map — Method

energy_map(coef, method)

Returns the Time Frequency Energy map or the Probability Density Energy map depending on the input method (TimeFrequency() or ProbabilityDensity()).

WaveletsExt.LDB.ldb — Method

ldb(X, y, wt[; dm=AsymmetricRelativeEntropy(), energy=TimeFrequency(),
    dp=BasisDiscriminantMeasure(), topk=size(X,1), m=size(X,1)])

The Local Discriminant Basis (LDB) feature selection algorithm developed by Saito and Coifman. This function takes in the input signals and their respective class labels, an orthogonal wavelet filter, and returns a set of expansion coefficients of reduced features and dimensions.

Arguments:

X::AbstractArray{<:Number,2}: the input set of signals. each column represents a signal.

y::AbstractVector{T}: the corresponding labels

wt::DiscreteWavelet: a discrete wavelet for transform purposes

dm::DiscriminantMeasure: the discriminant measure for the LDB algorithm. Supported measures are the AsymmetricRelativeEntropy(), LpEntropy() SymmetricRelativeEntropy(), and HellingerDistance(). Default is set to be AsymmetricRelativeEntropy().

energy::EnergyMap: the type of energy map used. Supported maps are TimeFrequency() and ProbabilityDensity(). Default is set to be TimeFrequency().

dp::DiscriminantPower(): the measure of discriminant power among expansion coefficients. Supported measures are BasisDiscriminantMeasure(), FishersClassSeparability(), and RobustFishersClassSeparability(). Default is set to be BasisDiscriminantMeasure().

topk::Integer: the top-k coefficients used in each node to determine the discriminant measure. Default is set to all coefficients.

m::Integer: the dimension of output after undergoing feature selection and transformation. Default is set to all coefficients of most discriminant subtree, ie. the length of each signals.

WaveletsExt.SIWPD.makesiwpdtree — Method

makesiwpdtree(n, L, d)

Returns the multi-level, multi-depth binary tree corresponding to the Shift- Invariant Wavelet Packet Decomposition.

WaveletsExt.SIWPD.siwpd — Method

siwpd(x, wt[, L=maxtransformlevels(x), d=L])

Computes the Shift-Invariant Wavelet Packet Decomposition originally developed by Cohen, Raz & Malah on the vector x using the discrete wavelet filter wt for L levels with depth d.

WaveletsExt.SWT.isdwt — Function

isdwt(xw, wt[, ε])

Performs the inverse stationary discrete wavelet transform (ISDWT) on the sdwt transform coefficients with respect to the Boolean Vector ε which represents the shifts to be used. If ε is not provided, the average-basis ISDWT will be computed instead.

See also: iswpt, sdwt

WaveletsExt.SWT.isdwt_step — Function

isdwt_step(v1, w1, j, s0, s1, g, h)

Perform one level of the inverse stationary discrete wavelet transform (ISDWT) on the vector v1 and w1, which is the j-th level scaling coefficients (Note the 0th level scaling coefficients is the raw signal). The vectors h and g are the detail and scaling filters.

Returns a vector v0 of the scaling and detail coefficients at level j-1.

See also: isdwt_step!

WaveletsExt.SWT.isdwt_step! — Function

isdwt_step!(v0, v1, w1, j, s0, s1, g, h)

Same as isdwt_step but without array allocation.

WaveletsExt.SWT.iswpt — Function

iswpt(xw, wt, ε[, L=maxtransformlevels(size(xw,1))])

iswpt(xw, wt, ε, tree)

iswpt(xw, wt[, L=maxtransformlevels(size(xw,1))])

iswpt(xw, wt, tree)

Performs the inverse stationary wavelet packet transform (ISWPT) on the swpd transform coefficients with respect to a given Boolean Vector that represents a binary tree and the BitVector ε which represents the shifts to be used. If ε is not provided, the average-basis ISWPT will be computed instead.

See also: isdwt, swpd

WaveletsExt.SWT.sdwt — Function

sdwt(x, wt[, L=maxtransformlevels(x)])

Perform a stationary discrete wavelet transform (SDWT) of the array x. The wavelet type wt determines the transform type and the wavelet class, see wavelet.

The number of transform levels L can be 1 ≤ L ≤ maxtransformlevels(x). Default value is set to maxtransformlevels(x).

Returns the n × (L+1) matrix (where n is the length of x) with the detail coefficients for level j in column (L-j+2). The scaling coefficients are in the 1st column.

See also: swpd, swpt, isdwt

WaveletsExt.SWT.sdwt_step — Function

sdwt_step(v, j, h, g)

Perform one level of the stationary discrete wavelet transform (SDWT) on the vector v, which is the j-th level scaling coefficients (Note the 0th level scaling coefficients is the raw signal). The vectors h and g are the detail and scaling filters.

Returns a tuple (v, w) of the scaling and detail coefficients at level j+1.

See also: sdwt_step!

WaveletsExt.SWT.sdwt_step! — Function

sdwt_step!(v1, w1, v, j, h, g)

Same as sdwt_step but without array allocation.

WaveletsExt.SWT.swpd — Function

swpd(x, wt[, L=maxtransformlevels(x)])

Perform a stationary wavelet packet decomposition (SPWD) of the array x. The wavelet type wt determines the transform type and the wavelet class, see wavelet.

The number of transform levels L can be 1 ≤ L ≤ maxtransformlevels(x). Default value is set to maxtransformlevels(x).

Returns the n × (2⁽ᴸ⁺¹⁾-1) matrix (where n is the length of x) with each column representing the nodes in the binary tree.

See also: swpt, sdwt, iswpt

WaveletsExt.SWT.swpt — Function

swpt(x, wt[, L=maxtransformlevels(x)])

swpt(x, wt, tree)

swpt(x, h, g, tree)

swpt(x, h, g, tree, i)

Performs the stationary wavelet packet transform (SWPT) of the vector x of length N = 2ᴸ. The wavelet type wt determines the transform type and the wavelet class, see wavelet.

The number of transform levels L can be 1 ≤ L ≤ maxtransformlevels(x). Default value is set to maxtransformlevels(x).

Returns the expansion coefficients of the SWPT of the size N × k. Each column represents a leaf node from tree. Number of returned columns can vary between 1 ≤ k ≤ N depending on the input tree.

See also: sdwt, swpd

WaveletsExt.Utils.coarsestscalingrange — Method

coarsestscalingrange(x, tree[, redundant=false])

coarsestscalingrange(n, tree[, redundant=false])

Given a binary tree, returns the index range of the coarsest scaling coefficients.

WaveletsExt.Utils.finestdetailrange — Method

finestdetailrange(x, tree[, redundant=false])

finestdetailrange(n, tree[, redundant=false])

Given a binary tree, returns the index range of the coarsest scaling coefficients.

WaveletsExt.Utils.generatesignals — Method

generatesignals(x, N, k[, noise=false, t=1])

Given a signal x, returns N shifted versions of the signal, each with shifts of multiples of k.

Setting noise = true allows randomly generated Gaussian noises of μ = 0, σ² = t to be added to the circularly shifted signals.

WaveletsExt.Utils.getleaf — Method

getleaf(tree)

Returns the leaf nodes of a tree.

WaveletsExt.Utils.left — Method

left(i)

Given the node index i, returns the index of its left node.

WaveletsExt.Utils.nodelength — Method

nodelength(N, L)

Returns the node length at level L of a signal of length N. Level L == 0 corresponds to the original input signal.

WaveletsExt.Utils.psnr — Method

psnr(x, x₀)

Returns the peak signal to noise ratio (PSNR) between original signal x₀ and noisy signal x.

WaveletsExt.Utils.relativenorm — Method

relativenorm(x, x₀[, p=2]) where T<:Number

Returns the relative norm of base p between original signal x₀ and noisy signal x.

WaveletsExt.Utils.right — Method

right(i)

Given the node index i, returns the index of its right node.

WaveletsExt.Utils.snr — Method

snr(x, x₀)

Returns the signal to noise ratio (SNR) between original signal x₀ and noisy signal x.

WaveletsExt.Utils.ssim — Method

ssim(x, x₀)

Wrapper for assess_ssim function from ImageQualityIndex.jl.

Returns the Structural Similarity Index Measure (SSIM) between the original signal/image x₀ and noisy signal/image x.

WaveletsExt.Visualizations.plot_tfbdry — Method

plot_tfbdry(tree[; start=0, nodecolor:white])

Given a tree, output a visual representation of the leaf nodes, user will have the option to start the node count of each level with 0 or 1.

WaveletsExt.Visualizations.treenodes_matrix — Method

treenodes_matrix(x)

Given a BitVector of nodes in a binary tree, output the matrix representation of the nodes.

WaveletsExt.Visualizations.wiggle! — Method

wiggle!(wav; taxis=1:size(wav,1), zaxis=1:size(wav,2), sc=1, 
    EdgeColor=:black, FaceColor=:black, Orient=:across, Overlap=true, 
    ZDir=:normal)

Plot a set of shaded wiggles on the current displayed graphics

Arguments

'wav::AbstractArray{<:Number,2}': matrix of waveform columns.

'taxis::AbstractVector=1:size(wav,1)': time axis vector

'zaxis::AbstractVector=1:size(wav,2)': space axis vector

'sc::Real=1': scale factor/magnification.

'EdgeColor::Symbol=:black': Sets edge of wiggles color.

'FaceColor::Symbol=:black': Sets shading color of wiggles.

'Overlap::Bool=true': How signals are scaled. true - Signals overlap (default); false - Signals are scaled so they do not overlap.

'Orient::Symbol=:across': Controls orientation of wiggles. :across - from left to right :down - from top to down

'ZDir::Symbol=:normal': Direction of space axis. :normal - First signal at bottom (default) :reverse - First signal at top.

Translated by Nicholas Hausch – MATLAB file provided by Naoki Saito The previous MATLAB version contributors are: Anthony K. Booer (SLB) and Bradley Marchand (NSWC-PC) Revised by Naoki Saito, Feb. 05, 2018

WaveletsExt.Visualizations.wiggle — Method

wiggle(wav; taxis=1:size(wav,1), zaxis=1:size(wav,2), sc=1, 
    EdgeColor=:black, FaceColor=:black, Orient=:across, Overlap=true, 
    ZDir=:normal)

Plot a set of shaded wiggles.

Arguments

'wav::AbstractArray{<:Number, 2}': matrix of waveform columns.

'taxis::AbstractVector=1:size(wav,1)': time axis vector

'zaxis::AbstractVector=1:size(wav,2)': space axis vector

'sc::Real=1': scale factor/magnification.

'EdgeColor::Symbol=:black': Sets edge of wiggles color.

'FaceColor::Symbol=:black': Sets shading color of wiggles.

'Overlap::Bool=true': How signals are scaled. true - Signals overlap (default); false - Signals are scaled so they do not overlap.

'Orient::Symbol=:across': Controls orientation of wiggles. :across - from left to right :down - from top to down

'ZDir::Symbol=:normal': Direction of space axis. :normal - First signal at bottom (default) :reverse - First signal at top.

Translated by Nicholas Hausch – MATLAB file provided by Naoki Saito The previous MATLAB version contributors are: Anthony K. Booer (SLB) and Bradley Marchand (NSWC-PC) Revised by Naoki Saito, Feb. 05, 2018

WaveletsExt.WPD.wpd — Function

wpd(x, wt[, L=maxtransformlevels(x)])

wpd(x, wt, hqf, gqf[, L=maxtransformlevels(x)])

Returns the wavelet packet decomposition WPD) for L levels for input signal(s) x.

WaveletsExt.WPD.wpd! — Method

wpd!(y, x, wt[, L=maxtransformlevels(x)])

wpd!(y, x, wt, hqf, gqf[, L=maxtransformlevels(x)])

Same as wpd but without array allocation.