Autocorrelation Wavelet Transform

WaveletsExt.ACWT.acwpt

WaveletsExt.ACWT.acwpt_step

WaveletsExt.ACWT.acwt

WaveletsExt.ACWT.acwt_step

WaveletsExt.ACWT.autocorr

WaveletsExt.ACWT.hacwt

WaveletsExt.ACWT.iacwpt

WaveletsExt.ACWT.iacwt

WaveletsExt.ACWT.iacwt!

WaveletsExt.ACWT.make_acqmfpair

WaveletsExt.ACWT.make_acreverseqmfpair

WaveletsExt.ACWT.pfilter

WaveletsExt.ACWT.qfilter

WaveletsExt.ACWT.vacwt

WaveletsExt.ACWT.acwpt — Function

acwpt(x, wt[, L=maxtransformlevels(x)])

Performs a discrete autocorrelation wavelet packet transform for a given signal x. The wavelet type wt determines the transform type. Refer to Wavelet.jl for a list of available methods.

Examples

acwpt(x, wavelet(WT.db4))

acwpt(x, wavelet(WT.db4), 4)

See also: acwt, acwpt_step, iacwpt

WaveletsExt.ACWT.acwpt_step — Function

acwpt_step(W, i, d, Qmf, Pmf)

Performs one level of the autocorrelation discrete wavelet packet transform (ACWPT) on the i-th node at depth d in the array W. The vectors Qmf and Pmf are the detail and scaling filters.

See also: acwpt, iacwpt

WaveletsExt.ACWT.acwt — Function

acwt(x, wt[, L=maxtransformlevels(x)])

acwt(x, wt[, Lrow=maxtransformlevels(x[1,:]), Lcol=maxtransformlevels(x[:,1])])

Performs a discrete autocorrelation wavelet transform for a given signal x. The signal can be 1D or 2D. The wavelet type wt determines the transform type. Refer to Wavelet.jl for a list of available methods.

Examples

acwt(x, wavelet(WT.db4))

acwt(x, wavelet(WT.db4), 4) # level 4 decomposition

See also: acwt_step, iacwt

WaveletsExt.ACWT.acwt_step — Function

acwt_step(v, j, h, g)

Performs one level of the autocorrelation discrete wavelet transform (ACWT) on the vector v, which is the j-th level scaling coefficients (Note the 0th level scaling coefficients is the raw signal). The vectors h and g are the detail and scaling filters.

Returns a tuple (v, w) of the scaling and detail coefficients at level j+1.

See also: acwt, iacwt

WaveletsExt.ACWT.autocorr — Method

autocorr(f::OrthoFilter)

Generates the autocorrelation filter for a given wavelet filter.

WaveletsExt.ACWT.hacwt — Function

hacwt(x, wt[, L=maxtransformlevels(x)])

Computes the column-wise discrete autocorrelation transform coeficients for 2D signals.

See also: vacwt

WaveletsExt.ACWT.iacwpt — Function

iacwpt(xw, tree, i)

Performs the inverse autocorrelation discrete wavelet packet transform, with respect to a decomposition tree.

See also: acwpt

WaveletsExt.ACWT.iacwt — Function

iacwt(xw::AbstractArray{<:Number,2})

iacwt(xw::AbstractArray{<:Number,4})

Performs the inverse autocorrelation discrete wavelet transform. Can be used for both the 1D and 2D case.

See also: iacwt!, acwt

WaveletsExt.ACWT.iacwt! — Function

iacwt!(xw::AbstractArray{<:Number,2})

Same as iacwt but performs the inverse autocorrelation discrete wavelet transform in place.

See also: iacwt, acwt

WaveletsExt.ACWT.make_acqmfpair — Method

make_acqmfpair(f::OrthoFilter)

Generates the autocorrelation quadratic mirror filters.

See also: make_acreverseqmfpair

WaveletsExt.ACWT.make_acreverseqmfpair — Method

make_acreverseqmfpair(f::OrthoFilter)

Generates the reverse autocorrelation quadratic mirror filters.

See also: make_acqmfpair

WaveletsExt.ACWT.pfilter — Method

pfilter(f::OrthoFilter)

Generates the high-pass autocorrelation filter

See also: qfilter

WaveletsExt.ACWT.qfilter — Method

qfilter(f::OrthoFilter)

Generates the low-pass autocorrelation filter.

See also: pfilter

WaveletsExt.ACWT.vacwt — Function

vacwt(x, wt[, L=maxtransformlevels(x)])

Computes the row-wise discrete autocorrelation transform coeficients for 2D signals.

See also: hacwt