Local Discriminant Basis

WaveletsExt.LDB.discriminant_measure

WaveletsExt.LDB.discriminant_power

WaveletsExt.LDB.energy_map

WaveletsExt.LDB.ldb

WaveletsExt.LDB.discriminant_measure — Method

discriminant_measure(Γ, dm)

Returns the discriminant measure calculated from the energy maps.

WaveletsExt.LDB.discriminant_power — Method

discriminant_power(coefs, tree, dp)

Returns the discriminant power of each leaf from the local discriminant basis (LDB) tree.

WaveletsExt.LDB.energy_map — Method

energy_map(coef, method)

Returns the Time Frequency Energy map or the Probability Density Energy map depending on the input method (TimeFrequency() or ProbabilityDensity()).

WaveletsExt.LDB.ldb — Method

ldb(X, y, wt[; dm=AsymmetricRelativeEntropy(), energy=TimeFrequency(),
    dp=BasisDiscriminantMeasure(), topk=size(X,1), m=size(X,1)])

The Local Discriminant Basis (LDB) feature selection algorithm developed by Saito and Coifman. This function takes in the input signals and their respective class labels, an orthogonal wavelet filter, and returns a set of expansion coefficients of reduced features and dimensions.

Arguments:

X::AbstractArray{<:Number,2}: the input set of signals. each column represents a signal.

y::AbstractVector{T}: the corresponding labels

wt::DiscreteWavelet: a discrete wavelet for transform purposes

dm::DiscriminantMeasure: the discriminant measure for the LDB algorithm. Supported measures are the AsymmetricRelativeEntropy(), LpEntropy() SymmetricRelativeEntropy(), and HellingerDistance(). Default is set to be AsymmetricRelativeEntropy().

energy::EnergyMap: the type of energy map used. Supported maps are TimeFrequency() and ProbabilityDensity(). Default is set to be TimeFrequency().

dp::DiscriminantPower(): the measure of discriminant power among expansion coefficients. Supported measures are BasisDiscriminantMeasure(), FishersClassSeparability(), and RobustFishersClassSeparability(). Default is set to be BasisDiscriminantMeasure().

topk::Integer: the top-k coefficients used in each node to determine the discriminant measure. Default is set to all coefficients.

m::Integer: the dimension of output after undergoing feature selection and transformation. Default is set to all coefficients of most discriminant subtree, ie. the length of each signals.