Translation-Invariant Wavelet Transform

WaveletsExt.TIWT.itidwt

WaveletsExt.TIWT.itidwt_step

WaveletsExt.TIWT.itidwt_step!

WaveletsExt.TIWT.itiwpd

WaveletsExt.TIWT.itiwpt

WaveletsExt.TIWT.tidwt

WaveletsExt.TIWT.tidwt_step

WaveletsExt.TIWT.tidwt_step!

WaveletsExt.TIWT.tiwpd

WaveletsExt.TIWT.tiwpt

Public API

WaveletsExt.TIWT.itidwt — Method

itidwt(xw, wt)

Computes the inverse translation-invariant discrete wavelet transform (iTIDWT) on xw.

Arguments

xw::AbstractArray{T,2} where T<:Number: TIDWT-transformed array.

wt::OrthoFilter: Orthogonal wavelet filter.

Returns

::Vector{T}: Inverse transformed signal.

Examples

using Wavelets, WaveletsExt

# Setup
x = generatesignals(:heavysine)
wt = wavelet(WT.haar)

# TIDWT
xw = tidwt(x, wt)

# iTIDWT
x̂ = itidwt(xw, wt)

See also: tidwt

WaveletsExt.TIWT.itiwpd — Method

itiwpd(xw, wt[, L])
itiwpd(xw, wt, tree)

Computes the inverse translation-invariant wavelet packet decomposition (iTIWPD) on xw.

Arguments

xw::AbstractArray{T,2} where T<:Number: TIDWT-transformed array.

wt::OrthoFilter: Orthogonal wavelet filter.

L::Integer: (Default: maxtransformlevels(size(xw,1))) Depth of inverse decomposition.

tree::BitVector: Representation of binary tree used as basis for the signal x.

Returns

::Vector{T}: Inverse transformed signal.

Examples

using Wavelets, WaveletsExt

# Setup
x = generatesignals(:heavysine)
wt = wavelet(WT.haar)

# TIWPD
xw = tiwpd(x, wt)

# iTIWPD
x̂ = itiwpd(xw, wt)

See also: tiwpd, itiwpt

WaveletsExt.TIWT.itiwpt — Method

itiwpt(xw, wt)

Computes the inverse translation-invariant wavelet packet transform (iTIWPT) on xw.

Arguments

xw::AbstractArray{T,2} where T<:Number: TIDWT-transformed array.

wt::OrthoFilter: Orthogonal wavelet filter.

Returns

::Vector{T}: Inverse transformed signal.

Examples

using Wavelets, WaveletsExt

# Setup
x = generatesignals(:heavysine)
wt = wavelet(WT.haar)

# TIWPT
xw = tiwpt(x, wt)

# iTIWPT
x̂ = itiwpt(xw, wt)

See also: tiwpt, itidwt

WaveletsExt.TIWT.tidwt — Method

tidwt(x, wt[, L])

Computes the translation-invariant discrete wavelet transform (TIDWT) for L levels.

Arguments

x::AbstractVector{T} where T<:Number: Original signal, preferably of size 2ᴷ where $K \in \mathbb{N}$.

wt::OrthoFilter: Orthogonal wavelet filter.

L::Integer: (Default: maxtransformlevels(x)) Number of levels of decomposition.

Returns

::Matrix{T}: Output from TIDWT on x.

Examples

using Wavelets, WaveletsExt

# Setup
x = generatesignals(:heavysine)
wt = wavelet(WT.haar)

# TIDWT
xw = tidwt(x, wt)

See also: itidwt

WaveletsExt.TIWT.tiwpd — Method

tiwpd(x, wt[, L])

Computes L levels of translation-invariant wavelet packet decomposition (TIWPD) on x.

Arguments

x::AbstractVector{T} where T<:Number: Original signal, preferably of size 2ᴷ where $K \in \mathbb{N}$.

wt::OrthoFilter: Orthogonal wavelet filter.

L::Integer: (Default: maxtransformlevels(x)) Number of levels of decomposition.

Returns

::Matrix{T}: Output from TIWPD on x.

Examples

using Wavelets, WaveletsExt

# Setup
x = generatesignals(:heavysine)
wt = wavelet(WT.haar)

# TIWPD
xw = tiwpd(x, wt)

See also: itiwpd, tiwpt

WaveletsExt.TIWT.tiwpt — Method

tiwpt(x, wt[, L])

Computes L levels of translation-invariant wavelet packet transform (TIWPT) on x.

Arguments

x::AbstractVector{T} where T<:Number: Original signal, preferably of size 2ᴷ where $K \in \mathbb{N}$.

wt::OrthoFilter: Orthogonal wavelet filter.

L::Integer: (Default: maxtransformlevels(x)) Number of levels of decomposition.

Returns

::Matrix{T}: Output from TIWPT on x.

Examples

using Wavelets, WaveletsExt

# Setup
x = generatesignals(:heavysine)
wt = wavelet(WT.haar)

# TIWPT
xw = tiwpt(x, wt)

See also: itiwpt, tidwt

Private API

WaveletsExt.TIWT.itidwt_step! — Method

itidwt_step!(v, w₁, w₂, d, h, g)

Same as itidwt_step but without array allocation.

Arguments

v::AbstractVector{T} where T<:Number: Output vector allocation.

w₁::AbstractVector{T} where T<:Number: d+1-level scaling coefficients.

w₂::AbstractVector{T} where T<:Number: d+1-level detail coefficients.

d::Integer: Depth level of output coefficients.

h::Vector{S} where S<:Number: High pass filter.

g::Vector{S} where S<:Number: Low pass filter.

Returns

v::AbstractVector{T}: Output from inverse translation-invariant wavelet transform of w₁ and w₂.

Examples

using WaveletsExt

# Setup
w₁, w₂ = randn(8), randn(8)
v = similar(w₁)
wt = wavelet(WT.haar)
g, h = WT.makereverseqmfpair(wt, false)

# One step of iTIDWT
tidwt_step(v, w₁, w₂, 0, h, g)

See also: itidwt_step

WaveletsExt.TIWT.itidwt_step — Method

itidwt_step(w₁, w₂, d, g, h)

Perform one level of the inverse translation-invariant discrete wavelet transform (ITIDWT) on the vector w₁ and w₂, which are the d+1-th level scaling and detail coefficients respectivaly (Note the 0-th level scaling coefficients correspond to the raw signal). The vectors h and g are the high and low pass filters.

Arguments

w₁::AbstractVector{T} where T<:Number: d+1-level scaling coefficients.

w₂::AbstractVector{T} where T<:Number: d+1-level detail coefficients.

d::Integer: Depth level of output coefficients.

h::Vector{S} where S<:Number: High pass filter.

g::Vector{S} where S<:Number: Low pass filter.

Returns

v::Vector{T}: Output of inverse translation-invariant wavelet transform.

Examples

using WaveletsExt

# Setup
w₁, w₂ = randn(8), randn(8)
wt = wavelet(WT.haar)
g, h = WT.makereverseqmfpair(wt, false)

# One step of iTIDWT
tidwt_step(w₁, w₂, 0, h, g)

See also: itidwt_step!

WaveletsExt.TIWT.tidwt_step! — Method

tidwt_step!(w₁, w₂, v, d, h, g)

Same as tidwt_step but without array allocation.

Arguments

w₁::AbstractVector{T} where T<:Number: Vector allocation for output from low pass filter.

w₂::AbstractVector{T} where T<:Number: Vector allocation for output from high pass filter.

v::AbstractVector{T} where T<:Number: Vector of coefficients from a node at level d.

d::Integer: Depth level of v.

h::Vector{S} where S<:Number: High pass filter.

g::Vector{S} where S<:Number: Low pass filter.

Returns

w₁::Vector{T}: Output from the low pass filter.

w₂::Vector{T}: Output from the high pass filter.

Examples

using Wavelets, WaveletsExt

# Setup
v = randn(8)
wt = wavelet(WT.haar)
g, h = WT.makereverseqmfpair(wt, true)
w₁ = zeros(8)
w₂ = zeros(8)

# One step of TIDWT
tidwt_step!(w₁, w₂, v, 0, h, g)

See also: tidwt_step

WaveletsExt.TIWT.tidwt_step — Method

tidwt_step(v, d, h, g)

Performs one level of the translation-invariant discrete wavelet transform (TWDWT) on the vector v, which is the d-th level scaling coefficients (Note the 0-th level scaling coefficients is the raw signal). The vectors h and g are the detail and scaling filters.

Arguments

v::AbstractVector{T} where T<:Number: Vector of coefficients from a node at level d.

d::Integer: Depth level of v.

h::Vector{S} where S<:Number: High pass filter.

g::Vector{S} where S<:Number: Low pass filter.

Returns

w₁::Vector{T}: Output from the low pass filter.

w₂::Vector{T}: Output from the high pass filter.

Examples

using Wavelets, WaveletsExt

# Setup
v = randn(8)
wt = wavelet(WT.haar)
g, h = WT.makereverseqmfpair(wt, true)

# One step of TIDWT
tidwt_step(v, 0, h, g)

See also: tidwt_step!