Utils

WaveletsExt.Utils.ClassData

WaveletsExt.Utils.coarsestscalingrange

WaveletsExt.Utils.duplicatesignals

WaveletsExt.Utils.finestdetailrange

WaveletsExt.Utils.generateclassdata

WaveletsExt.Utils.generatesignals

WaveletsExt.Utils.getcolrange

WaveletsExt.Utils.getleaf

WaveletsExt.Utils.getquadtreelevel

WaveletsExt.Utils.getrowrange

WaveletsExt.Utils.left

WaveletsExt.Utils.main2depthshift

WaveletsExt.Utils.makequadtree

WaveletsExt.Utils.nodelength

WaveletsExt.Utils.packet

WaveletsExt.Utils.psnr

WaveletsExt.Utils.relativenorm

WaveletsExt.Utils.right

WaveletsExt.Utils.snr

WaveletsExt.Utils.ssim

Public API

WaveletsExt.Utils.ClassData — Type

ClassData(type, s₁, s₂, s₃)

Based on the input type, generates 3 classes of signals with sample sizes s₁, s₂, and s₃ respectively. Accepted input types are:

:tri: Triangular signals of length 32

:cbf: Cylinder-Bell-Funnel signals of length 128

Based on N. Saito and R. Coifman in "Local Discriminant Basis and their Applications" in the Journal of Mathematical Imaging and Vision, Vol. 5, 337-358 (1995).

See also: generateclassdata

WaveletsExt.Utils.coarsestscalingrange — Method

coarsestscalingrange(x, tree[, redundant])
coarsestscalingrange(n, tree[, redundant])

Given a binary tree, returns the index range of the coarsest scaling coefficients.

Arguments

x::AbstractArray{T} where T<:Number: Decomposed 1D-signal.

n::Integer: Length of signal of interest.

tree::BitVector: Binary tree.

redundant::Bool: (Default: false) Whether the wavelet decomposition is redundant. Examples of redundant wavelet transforms are the Autocorrelation wavelet transform (ACWT), Stationary wavelet transform (SWT), and the Maximal Overlap wavelet transform (MOWT).

Returns

UnitRange{Integer} or ::Tuple{UnitRange{Integer}, Integer}: The index range of the coarsest scaling subspace based on the input binary tree.

Examples

using Wavelets, WaveletsExt

x = randn(8)
wt = wavelet(WT.haar)
tree = maketree(x)

# Non-redundant wavelet transform
xw = wpd(x, wt)
coarsestscalingrange(xw, tree)          # 1:1

# Redundant wavelet transform
xw = swpd(x, wt)
coarsestscalingrange(xw, tree, true)    # (1:8, 8)

See also: finestdetailrange

WaveletsExt.Utils.duplicatesignals — Method

duplicatesignals(x, N, k[, noise=false, t=1])

Given a signal x, returns N shifted versions of the signal, each with shifts of multiples of k.

Setting noise = true allows randomly generated Gaussian noises of μ = 0, σ² = t to be added to the circularly shifted signals.

Arguments

x::AbstractVector{T} where T<:Number: 1D-signal to be duplicated.

N::Integer:: Number of duplicated signals.

k::Integer:: Circular shift size for each duplicated signal.

noise::Bool: (Default: false) Whether or not to add Gaussian noise.

t::Real: (Default: 1) Relative size of noise.

Returns

::Array{T}: Duplicated signals.

Examples

using WaveletsExt

x = generatesignals(:blocks)
duplicatesignals(x, 5, 0)      # [x x x x x]

See also: generatesignals

WaveletsExt.Utils.finestdetailrange — Method

finestdetailrange(x, tree[, redundant])
finestdetailrange(n, tree[, redundant])

Given a binary tree, returns the index range of the finest detail coefficients.

Arguments

x::AbstractArray{T} where T<:Number: Decomposed 1D-signal.

n::Integer: Length of signal of interest.

tree::BitVector: Binary tree.

redundant::Bool: (Default: false) Whether the wavelet decomposition is redundant. Examples of redundant wavelet transforms are the Autocorrelation wavelet transform (ACWT), Stationary wavelet transform (SWT), and the Maximal Overlap wavelet transform (MOWT).

Returns

UnitRange{Integer} or ::Tuple{UnitRange{Integer}, Integer}: The index range of the finest detail subspace based on the input binary tree.

Examples

using Wavelets, WaveletsExt

x = randn(8)
wt = wavelet(WT.haar)
tree = maketree(x)

# Non-redundant wavelet transform
xw = wpd(x, wt)
finestdetailrange(xw, tree)          # 8:8

# Redundant wavelet transform
xw = swpd(x, wt)
finestdetailrange(xw, tree, true)    # (1:8, 15)

See also: coarsestscalingrange

WaveletsExt.Utils.generateclassdata — Function

generateclassdata(c[, shuffle])

Generates 3 classes of data given a ClassData struct as an input. Returns a matrix containing the 3 classes of signals and a vector containing their corresponding labels.

Based on N. Saito and R. Coifman in "Local Discriminant Basis and their Applications" in the Journal of Mathematical Imaging and Vision, Vol. 5, 337-358 (1995).

Arguments

c::ClassData: Type of signal classes to generate.

shuffle::Bool: (Default: true). Whether or not to shuffle the signals.

Returns

::Matrix{Float64}: Generated signals.

::Vector{Int64}: Class corresponding to each column of generated signals.

Examples

using WaveletsExt

c = ClassData(:tri, 100, 100, 100)
generateclassdata(c)

See also: ClassData

WaveletsExt.Utils.generatesignals — Function

generatesignals(fn, L)

Generates a signal of length 2ᴸ given the function symbol fn. Current accepted inputs below are based on D. Donoho and I. Johnstone in "Adapting to Unknown Smoothness via Wavelet Shrinkage" Preprint Stanford, January 93, p 27-28.

:blocks

:bumps

:heavysine

:doppler

:quadchirp

:mishmash

The code for this function is adapted and translated based on MATLAB's Wavelet Toolbox's wnoise function.

Arguments

fn::Symbol: Type of function/signal to generate.

L::Integer: (Default = 7) Size of the signal to generate. Will return a signal of size 2ᴸ.

Returns

::Vector{Float64}: Signal of length 2ᴸ.

Examples

using WaveletsExt

generatesignals(:bumps, 8)

WaveletsExt.Utils.getcolrange — Method

getcolrange(n, idx)

Get the column range from a matrix with n columns that corresponds to idx from a quadtree.

Arguments

n::Integer: Number of columns in matrix.

idx::T where T<:Integer: Index from a quadtree corresponding to the matrix.

Returns

::UnitRange{Int64}: Column range in matrix that corresponds to idx.

Examples

using WaveletsExt

x = randn(8,8)
tree = makequadtree(x, 3, :full)
getcolrange(8,3)            # 5:8

See also: makequadtree, getrowrange

WaveletsExt.Utils.getleaf — Method

getleaf(tree)

Returns the leaf nodes of a tree.

Arguments

tree::BitVector: BitVector to represent binary tree.

Returns

::BitVector: BitVector that can represent a binary tree, but only the leaves are labeled 1, the rest of the nodes are labeled 0.

Examples

using Wavelets, WaveletsExt

tree = maketree(4, 2, :dwt)     # [1,1,0]
getleaf(tree)                   # [0,0,1,1,1,0,0]

WaveletsExt.Utils.getquadtreelevel — Method

getquadtreelevel(idx)

Get level of idx in the quadtree.

Arguments

idx::T where T<:Integer: Index of a quadtree.

Returns

::T: Level of idx.

Examples

using WaveletsExt

getquadtreelevel(1)     # 0
getquadtreelevel(3)     # 1

See also: makequadtree

WaveletsExt.Utils.getrowrange — Method

getrowrange(n, idx)

Get the row range from a matrix with n rows that corresponds to idx from a quadtree.

Arguments

n::Integer: Number of rows in matrix.

idx::T where T<:Integer: Index from a quadtree corresponding to the matrix.

Returns

::UnitRange{Int64}: Row range in matrix that corresponds to idx.

Examples

using WaveletsExt

x = randn(8,8)
tree = makequadtree(x, 3, :full)
getrowrange(8,3)            # 1:4

See also: makequadtree, getcolrange

WaveletsExt.Utils.left — Method

left(i)

Given the node index i, returns the index of its left node.

Arguments

i::Integer: Index of the node of interest.

Returns

::Integer: Index of left child.

Examples

using Wavelets, WaveletsExt

left(3)     # 6

See also: right

WaveletsExt.Utils.makequadtree — Method

makequadtree(x, L[, s])

Build quadtree for 2D wavelet transform. Indexing of the quadtree are as follows:

Level 0                 Level 1                 Level 2             ...
_________________       _________________       _________________
|               |       |   2   |   3   |       |_6_|_7_|10_|11_|
|       1       |       |_______|_______|       |_8_|_9_|12_|13_|   ...
|               |       |   4   |   5   |       |14_|15_|18_|19_|
|_______________|       |_______|_______|       |16_|17_|20_|21_|

Arguments

x::AbstractArray{T,2} where T<:Number: Input array.

L::Integer: Number of decomposition levels.

s::Symbol: (Default: :full) Type of quadtree. Available types are :full and :dwt.

Returns

::BitVector: Quadtree representation.

Examples

using WaveletsExt

x = randn(16,16)
makequadtree(x, 3)

WaveletsExt.Utils.nodelength — Method

nodelength(N, L)

Returns the node length at level L of a signal of length N when performaing wavelet packet decomposition. Level L == 0 corresponds to the original input signal.

Arguments

N::Integer: Length of signal.

L::Integer: Level of signal.

Returns

::Integer: Length of nodes at level L.

WaveletsExt.Utils.psnr — Method

psnr(x, x₀)

Returns the peak signal to noise ratio (PSNR) between original signal x₀ and noisy signal x.

PSNR definition: $10 \log_{10} \frac{\max{(x_0)}^2}{MSE(x, x_0)}$

Arguments

x::AbstractVector{T} where T<:Number: Signal with noise.

x₀::AbstractVector{T} where T<:Number: Reference signal.

Returns

::AbstractFloat: The PSNR between x and x₀.

Examples

using WaveletsExt

x = randn(8)
y = randn(8)

psnr(x, y)

See also: relativenorm, snr, ssim

WaveletsExt.Utils.relativenorm — Method

relativenorm(x, x₀[, p]) where T<:Number

Returns the relative norm of base p between original signal x₀ and noisy signal x.

Arguments

x::AbstractVector{T} where T<:Number: Signal with noise.

x₀::AbstractVector{T} where T<:Number: Reference signal.

p::Real: (Default: 2) $p$-norm to be computed.

Returns

::AbstractFloat: The relative norm between x and x₀.

Examples

using WaveletsExt

x = randn(8)
y = randn(8)

relativenorm(x, y)

See also: psnr, snr, ssim

WaveletsExt.Utils.right — Method

right(i)

Given the node index i, returns the index of its right node.

Arguments

i::Integer: Index of the node of interest.

Returns

::Integer: Index of right child.

Examples

using Wavelets, WaveletsExt

left(3)     # 7

See also: left

WaveletsExt.Utils.snr — Method

snr(x, x₀)

Returns the signal to noise ratio (SNR) between original signal x₀ and noisy signal x.

SNR definition: $20 \log_{10} \frac{||x_0||_2}{||x-x_0||_2}$

Arguments

x::AbstractVector{T} where T<:Number: Signal with noise.

x₀::AbstractVector{T} where T<:Number: Reference signal.

Returns

::AbstractFloat: The SNR between x and x₀.

Examples

using WaveletsExt

x = randn(8)
y = randn(8)

snr(x, y)

See also: relativenorm, psnr, ssim

WaveletsExt.Utils.ssim — Method

ssim(x, x₀)

Wrapper for assess_ssim function from ImageQualityIndex.jl.

Returns the Structural Similarity Index Measure (SSIM) between the original signal/image x₀ and noisy signal/image x.

Arguments

x::AbstractArray{T} where T<:Number: Signal with noise.

x₀::AbstractArray{T} where T<:Number: Reference signal.

Returns

::AbstractFloat: The SNR between x and x₀.

Examples

using WaveletsExt

x = randn(8)
y = randn(8)

ssim(x, y)

See also: relativenorm, psnr, snr

Private API

Wavelets.Util.maxtransformlevels — Method

maxtransformlevels(n)

maxtransformlevels(x[, dims])

Extension function from Wavelets.jl. Finds the max number of transform levels for an array x. If dims is provided for a multidimensional array, then it finds the max transform level for that corresponding dimension.

Arguments

x::AbstractArray: Input array.

dims::Integer: Dimension used for wavelet transform.

n::Integer: Length of input vector.

Returns

::Integer: Max number of transform levels.

Examples

using Wavelets, WaveletsExt

# Define random signal
x = randn(64, 128)

# Max transform levels and their corresponding return values
maxtransformlevels(128)     # 7
maxtransformlevels(x)       # 6
maxtransformlevels(x, 2)    # 7

WaveletsExt.Utils.main2depthshift — Method

main2depthshift(sm, L)

Given the overall shift sm, compute the cumulative shift at each depth. Useful for computing the shift based inverse redundant wavelet transforms.

Arguments

sm::Integer: Overall shift.

L::Integer: The total number of depth.

Returns

::Vector{Int}: Vector where each entry i describes the shift at depth i-1.

Examples

using WaveletsExt

Utils.main2depthshift(10, 4)      # [0, 0, 2, 2, 10]
Utils.main2depthshift(5, 5)       # [0, 1, 1, 5, 5, 5]

WaveletsExt.Utils.packet — Method

packet(d, b, n)

Packet table indexing.

Arguments

d::Integer: Depth of splitting in packet decomposition. Note: Depth of root node is 0.

b::Integer: Block index among 2ᵈ possibilities at depth d. Note: Block indexing starts from 0.

n::Integer: Length of signal.

Returns

::UnitRange{Int64}: Index range of block b of signal of length n at level d.

Examples

using WaveletsExt

Utils.packet(0, 0, 8)       # 1:8
Utils.packet(2, 1, 8)       # 3:4

Translated from Wavelab850 by Zeng Fung Liew.