Visualizations

WaveletsExt.Visualizations.plot_tfbdry

WaveletsExt.Visualizations.treenodes_matrix

WaveletsExt.Visualizations.wiggle

WaveletsExt.Visualizations.wiggle!

WaveletsExt.Visualizations.plot_tfbdry — Method

plot_tfbdry(tree[; start=0, nodecolor:white])

Given a tree, output a visual representation of the leaf nodes, user will have the option to start the node count of each level with 0 or 1.

WaveletsExt.Visualizations.treenodes_matrix — Method

treenodes_matrix(x)

Given a BitVector of nodes in a binary tree, output the matrix representation of the nodes.

WaveletsExt.Visualizations.wiggle! — Method

wiggle!(wav; taxis=1:size(wav,1), zaxis=1:size(wav,2), sc=1, 
    EdgeColor=:black, FaceColor=:black, Orient=:across, Overlap=true, 
    ZDir=:normal)

Plot a set of shaded wiggles on the current displayed graphics

Arguments

wav::AbstractArray{<:Number,2}: matrix of waveform columns.

taxis::AbstractVector=1:size(wav,1): time axis vector

zaxis::AbstractVector=1:size(wav,2): space axis vector

sc::Real=1: scale factor/magnification.

EdgeColor::Symbol=:black: Sets edge of wiggles color.

FaceColor::Symbol=:black: Sets shading color of wiggles.

Overlap::Bool=true: How signals are scaled.

true - Signals overlap (default);

false - Signals are scaled so they do not overlap.

Orient::Symbol=:across: Controls orientation of wiggles.

:across - from left to right

:down - from top to down

ZDir::Symbol=:normal: Direction of space axis.

:normal - First signal at bottom (default)

:reverse - First signal at top.

Translated by Nicholas Hausch – MATLAB file provided by Naoki Saito

The previous MATLAB version contributors are: Anthony K. Booer (SLB) and Bradley Marchand (NSWC-PC)

Revised by Naoki Saito, Feb. 05, 2018

Maintained by Zeng Fung Liew for newest Julia version compatibility.

See also: wiggle

WaveletsExt.Visualizations.wiggle — Method

wiggle(wav; taxis=1:size(wav,1), zaxis=1:size(wav,2), sc=1, 
    EdgeColor=:black, FaceColor=:black, Orient=:across, Overlap=true, 
    ZDir=:normal)

Plot a set of shaded wiggles.

Arguments

wav::AbstractArray{<:Number, 2}: matrix of waveform columns.

taxis::AbstractVector=1:size(wav,1): time axis vector

zaxis::AbstractVector=1:size(wav,2): space axis vector

sc::Real=1: scale factor/magnification.

EdgeColor::Symbol=:black: Sets edge of wiggles color.

FaceColor::Symbol=:black: Sets shading color of wiggles.

Overlap::Bool=true: How signals are scaled.

true - Signals overlap (default);

false - Signals are scaled so they do not overlap.

Orient::Symbol=:across: Controls orientation of wiggles.

:across - from left to right

:down - from top to down

ZDir::Symbol=:normal: Direction of space axis.

:normal - First signal at bottom (default)

:reverse - First signal at top.

Translated by Nicholas Hausch – MATLAB file provided by Naoki Saito

The previous MATLAB version contributors are: Anthony K. Booer (SLB) and Bradley Marchand (NSWC-PC)

Revised by Naoki Saito, Feb. 05, 2018

Maintained by Zeng Fung Liew for newest Julia version compatibility.

See also: wiggle!