Wavelet Packet Decomposition

WaveletsExt.WPD.dwtall

WaveletsExt.WPD.wpd

WaveletsExt.WPD.wpd!

WaveletsExt.WPD.wpdall

WaveletsExt.WPD.wptall

Public API

WaveletsExt.WPD.dwtall — Method

dwtall(x, wt[, L])

Computes the discrete wavelet transform (DWT) on each slice of signal. Signals are sliced on the $n$-th dimension for an $n$-dimensional input x.

Note: dwt is currently available for 1-D, 2-D, and 3-D signals only.

Arguments

x::AbstractArray{T} where T<:Number: Input signals, where each slice corresponds to one signal. For a set of input signals x of dimension $n$, signals are sliced on the $n$-th dimension.

wt::OrthoFilter: Wavelet used.

L::Integer: (Default: Wavelets.maxtransformlevels(xᵢ)) Number of levels of wavelet transforms.

Returns

::Array{T}: Slices of transformed signals. Signals are sliced the same way as the input signal x.

Examples

using Wavelets, WaveletsExt

# Generate random signals
x = randn(32, 5)
# Create wavelet
wt = wavelet(WT.db4)

# DWT on all signals in x
xw = dwtall(x, wt)

See also: wpdall, wptall

WaveletsExt.WPD.wpd! — Method

wpd!(y, x, wt[, L])

wpd!(y, x, wt[, L; standard])

Same as wpd but without array allocation.

Arguments

y::AbstractArray{T,2} where T<:Number or y::AbstractArray{T,3} where T<:Number: An allocated array to write the outputs of x onto.

x::AbstractVector{T} where T<:Number or x::AbstractArray{T,2} where T<:Number: Input vector/matrix. A vector input undergoes 1D wavelet decomposition whereas a matrix input undergoes 2D wavelet decomposition.

wt::OrthoFilter: Wavelet filter.

L::Integer: (Default: maxtransformlevels(x)) Number of levels for wavelet decomposition.

standard::Bool: (Default: true) Whether to compute the standard or non-standard wavelet transform. Only applicable for 2D signals.

Returns

y::AbstractArray{T,2} where T<:Number or y::AbstractArray{T,3} where T<:Number: Decomposed signal. For an input vector, output is a 2D matrix where each column corresponds to a level of decomposition. For an input matrix, output is a 3D array where each slice of dimension 3 corresponds to a level of decomposition.

Examples:

using Wavelets, WaveletsExt

# 1D wavelet decomposition
x = randn(8)
xw = Array{eltype(x)}(undef, (8,3))
wt = wavelet(WT.haar)
wpd!(xw, x, wt)

# 2D wavelet decomposition
x = randn(8,8)
xw = Array{eltype(x)}(undef, (8,8,3))
wt = wavelet(WT.haar)
wpd!(xw, x, wt)

See also: wpd

WaveletsExt.WPD.wpd — Method

wpd(x, wt[, L])

wpd(x, wt[, L; standard])

Returns the wavelet packet decomposition (WPD) for L levels for input signal x.

Arguments

x::AbstractVector{T} where T<:Number or x::AbstractArray{T,2} where T<:Number: Input vector/matrix. A vector input undergoes 1D wavelet decomposition whereas a matrix input undergoes 2D wavelet decomposition.

wt::OrthoFilter: Wavelet filter.

L::Integer: (Default: maxtransformlevels(x)) Number of levels for wavelet decomposition.

standard::Bool: (Default: true) Whether to compute the standard or non-standard wavelet transform. Only applicable for 2D signals.

Returns

::Array{T,2} or ::Array{T,3}: Decomposed signal. For an input vector, output is a 2D matrix where each column corresponds to a level of decomposition. For an input matrix, output is a 3D array where each slice of dimension 3 corresponds to a level of decomposition.

Examples:

using Wavelets, WaveletsExt

# 1D wavelet decomposition
x = randn(8)
wt = wavelet(WT.haar)
xw = wpd(x, wt)

# 2D wavelet decomposition
x = randn(8,8)
wt = wavelet(WT.haar)
xw = wpd(x, wt)

See also: wpd!

WaveletsExt.WPD.wpdall — Method

wpdall(x, wt[, L])

wpdall(x, wt[, L; standard])

Computes the wavelet packet decomposition (WPD) on each slice of signal. Signals are sliced on the $n$-th dimension for an $n$-dimensional input x.

Note: wpd is currently available for 1-D signals only.

Arguments

x::AbstractArray{T} where T<:Number: Input signals, where each slice corresponds to one signal. For a set of input signals x of dimension $n$, signals are sliced on the $n$-th dimension.

wt::OrthoFilter: Wavelet used.

L::Integer: (Default: Wavelets.maxtransformlevels(xᵢ)) Number of levels of wavelet decomposition.

standard::Bool: (Default: true) Whether to compute the standard or non-standard wavelet transform. Only applicable for 2D signals.

Returns

::Array{T}: Array of decomposed signals. Signals are sliced by the final dimension.

Examples

using Wavelets, WaveletsExt

# Generate random signals
x = randn(32, 5)
# Create wavelet
wt = wavelet(WT.db4)

# WPT on all signals in x
xw = wpdall(x, wt)

See also: wptall, dwtall, wpd, wpd!

WaveletsExt.WPD.wptall — Method

wptall(x, wt[, L])

wptall(x, wt, tree)

wptall(x, wt[, L; standard])

wptall(x, wt, tree[; standard])

Computes the wavelet packet transform (WPT) on each slice of signal. Signals are sliced on the $n$-th dimension for an $n$-dimensional input x.

Note: wpt is currently available for 1-D signals only.

Arguments

x::AbstractArray{T} where T<:Number: Input signals, where each slice corresponds to one signal. For a set of input signals x of dimension $n$, signals are sliced on the $n$-th dimension.

wt::OrthoFilter: Wavelet used.

L::Integer: (Default: Wavelets.maxtransformlevels(xᵢ)) Number of levels of wavelet decomposition.

tree::BitVector: (Default: Wavelets.maketree(xᵢ, :full)) Tree to follow for wavelet decomposition. Default value is only applicable for 1D signals.

standard::Bool: (Default: true) Whether to compute the standard or non-standard wavelet transform. Only applicable for 2D signals.

Returns

::Array{T}: Slices of transformed signals. Signals are sliced the same way as the input signal x.

Examples

using Wavelets, WaveletsExt

# Generate random signals
x = randn(32, 5)
# Create wavelet
wt = wavelet(WT.db4)

# WPT on all signals in x
xw = wptall(x, wt)

See also: wpdall, dwtall, wpt

Private API

Wavelets.Transforms.wpt! — Method

wpt!(y, x, wt[, L])

wpt!(y, x, wt, tree)

wpt!(y, x, wt[, L; standard])

wpt!(y, x, wt, tree[; standard])

Same as wpt but without array allocation.

Arguments

x::AbstractVector{T} where T<:Number or x::AbstractArray{T,2} where T<:Number: Input vector/matrix. A vector input undergoes 1D wavelet decomposition whereas a matrix input undergoes 2D wavelet decomposition.

wt::OrthoFilter: Wavelet filter.

L::Integer: (Default: maxtransformlevels(x)) Number of levels for wavelet decomposition.

tree::BitVector: Quadtree to transform to be computed accordingly.

standard::Bool: (Default: true) Whether to compute the standard or non-standard wavelet transform. Only applicable for 2D signals.

Returns

::Array{T,1} or ::Array{T,2}: Transformed signal.

Examples:

using Wavelets, WaveletsExt

# 1D wavelet decomposition
x = randn(8)
xw = similar(x)
wt = wavelet(WT.haar)
wpt!(xw, x, wt)

# 2D wavelet decomposition
x = randn(8,8)
xw = similar(x)
wt = wavelet(WT.haar)
wpt!(xw, x, wt)

See also: wpt, makequadtree

Wavelets.Transforms.wpt — Method

wpt(x, wt[, L])

wpt(x, wt, tree)

wpt(x, wt[, L; standard])

wpt(x, wt, tree[; standard])

Returns the wavelet packet transform (WPT) for L levels or by given quadratic tree.

Arguments

x::AbstractVector{T} where T<:Number or x::AbstractArray{T,2} where T<:Number: Input vector/matrix. A vector input undergoes 1D wavelet decomposition whereas a matrix input undergoes 2D wavelet decomposition.

wt::OrthoFilter: Wavelet filter.

L::Integer: (Default: maxtransformlevels(x)) Number of levels for wavelet decomposition.

tree::BitVector: Quadtree to transform to be computed accordingly.

standard::Bool: (Default: true) Whether to compute the standard or non-standard wavelet transform. Only applicable for 2D signals.

Returns

::Array{T,1} or ::Array{T,2}: Transformed signal.

Examples:

using Wavelets, WaveletsExt

# 1D wavelet decomposition
x = randn(8)
wt = wavelet(WT.haar)
xw = wpt(x, wt)

# 2D wavelet decomposition
x = randn(8,8)
wt = wavelet(WT.haar)
xw = wpt(x, wt)

See also: wpt!, makequadtree