Autocorrelation Wavelet Transform

WaveletsExt.ACWT.acdwt

WaveletsExt.ACWT.acdwt_step

WaveletsExt.ACWT.acwpd

WaveletsExt.ACWT.acwpt_step

WaveletsExt.ACWT.autocorr

WaveletsExt.ACWT.hacdwt

WaveletsExt.ACWT.iacdwt

WaveletsExt.ACWT.iacdwt!

WaveletsExt.ACWT.iacwpt

WaveletsExt.ACWT.make_acqmfpair

WaveletsExt.ACWT.make_acreverseqmfpair

WaveletsExt.ACWT.pfilter

WaveletsExt.ACWT.qfilter

WaveletsExt.ACWT.vacdwt

Public API

WaveletsExt.ACWT.acdwt — Function

acdwt(x, wt[, L=maxtransformlevels(x)])

acdwt(x, wt[, Lrow=maxtransformlevels(x[1,:]), Lcol=maxtransformlevels(x[:,1])])

Performs a discrete autocorrelation wavelet transform for a given signal x. The signal can be 1D or 2D. The wavelet type wt determines the transform type. Refer to Wavelet.jl for a list of available methods.

Examples

acdwt(x, wavelet(WT.db4))

acdwt(x, wavelet(WT.db4), 4) # level 4 decomposition

See also: acdwt_step, iacdwt

WaveletsExt.ACWT.acwpd — Function

acwpd(x, wt[, L=maxtransformlevels(x)])

Performs a discrete autocorrelation wavelet packet transform for a given signal x. The wavelet type wt determines the transform type. Refer to Wavelet.jl for a list of available methods.

Examples

acwpd(x, wavelet(WT.db4))

acwpd(x, wavelet(WT.db4), 4)

See also: acdwt, acwpt_step, iacwpt

WaveletsExt.ACWT.iacdwt — Function

iacdwt(xw::AbstractArray{<:Number,2})

iacdwt(xw::AbstractArray{<:Number,4})

Performs the inverse autocorrelation discrete wavelet transform. Can be used for both the 1D and 2D case.

See also: iacdwt!, acdwt

WaveletsExt.ACWT.iacwpt — Function

iacwpd(xw, tree, i)

Performs the inverse autocorrelation discrete wavelet packet transform, with respect to a decomposition tree.

See also: acwpd

Private API

WaveletsExt.ACWT.acdwt_step — Function

acdwt_step(v, j, h, g)

Performs one level of the autocorrelation discrete wavelet transform (acdwt) on the vector v, which is the j-th level scaling coefficients (Note the 0th level scaling coefficients is the raw signal). The vectors h and g are the detail and scaling filters.

Returns a tuple (v, w) of the scaling and detail coefficients at level j+1.

See also: acdwt, iacdwt

WaveletsExt.ACWT.acwpt_step — Function

acwpt_step(W, i, d, Qmf, Pmf)

Performs one level of the autocorrelation discrete wavelet packet transform (ACWPT) on the i-th node at depth d in the array W. The vectors Qmf and Pmf are the detail and scaling filters.

See also: acwpd, iacwpt

WaveletsExt.ACWT.autocorr — Method

autocorr(f::OrthoFilter)

Generates the autocorrelation filter for a given wavelet filter.

WaveletsExt.ACWT.hacdwt — Method

hacdwt(x, wt[, L=maxtransformlevels(x,2)])

Computes the column-wise discrete autocorrelation transform coeficients for 2D signals.

See also: vacdwt

WaveletsExt.ACWT.iacdwt! — Function

iacdwt!(xw::AbstractArray{<:Number,2})

Same as iacdwt but performs the inverse autocorrelation discrete wavelet transform in place.

See also: iacdwt, acdwt

WaveletsExt.ACWT.make_acqmfpair — Method

make_acqmfpair(f::OrthoFilter)

Generates the autocorrelation quadratic mirror filters.

See also: make_acreverseqmfpair

WaveletsExt.ACWT.make_acreverseqmfpair — Method

make_acreverseqmfpair(f::OrthoFilter)

Generates the reverse autocorrelation quadratic mirror filters.

See also: make_acqmfpair

WaveletsExt.ACWT.pfilter — Method

pfilter(f::OrthoFilter)

Generates the high-pass autocorrelation filter

See also: qfilter

WaveletsExt.ACWT.qfilter — Method

qfilter(f::OrthoFilter)

Generates the low-pass autocorrelation filter.

See also: pfilter

WaveletsExt.ACWT.vacdwt — Method

vacdwt(x, wt[, L=maxtransformlevels(x)])

Computes the row-wise discrete autocorrelation transform coeficients for 2D signals.

See also: hacdwt