Best Basis

WaveletsExt.BestBasis.BB

WaveletsExt.BestBasis.BBCost

WaveletsExt.BestBasis.BestBasisType

WaveletsExt.BestBasis.CostFunction

WaveletsExt.BestBasis.DifferentialEntropyCost

WaveletsExt.BestBasis.JBB

WaveletsExt.BestBasis.JBBCost

WaveletsExt.BestBasis.LSDB

WaveletsExt.BestBasis.LSDBCost

WaveletsExt.BestBasis.LogEnergyEntropyCost

WaveletsExt.BestBasis.LoglpCost

WaveletsExt.BestBasis.NormCost

WaveletsExt.BestBasis.SIBB

WaveletsExt.BestBasis.ShannonEntropyCost

WaveletsExt.BestBasis.bestbasis_treeselection

WaveletsExt.BestBasis.bestbasiscoef

WaveletsExt.BestBasis.tree_costs

WaveletsExt.BestBasis.BB — Type

BB([; cost=LoglpCost(2), redundant=false])

Best Basis (BB). Set redundant=true when running BB with redundant wavelet transforms such as SWT or ACWT.

See also: BestBasisType, LSDB, JBB, SIBB

WaveletsExt.BestBasis.BBCost — Type

BBCost <: CostFunction

Cost function abstract type specifically for BB.

See also: CostFunction, LSDBCost, JBBCost

WaveletsExt.BestBasis.BestBasisType — Type

Abstract type for best basis. Current available types are:

LSDB

JBB

BB

SIBB

WaveletsExt.BestBasis.CostFunction — Type

Cost function abstract type.

See also: LSDBCost, JBBCost, BBCost

WaveletsExt.BestBasis.DifferentialEntropyCost — Type

DifferentialEntropyCost <: LSDBCost

Differential entropy cost used for LSDB.

See also: CostFunction, LSDBCost

WaveletsExt.BestBasis.JBB — Type

JBB([; cost=LoglpCost(2), redundant=false])

Joint Best Basis (JBB). Set redundant=true when running JBB with redundant wavelet transforms such as SWT or ACWT.

See also: BestBasisType, LSDB, BB, SIBB

WaveletsExt.BestBasis.JBBCost — Type

JBBCost <: CostFunction

Cost function abstract type specifically for JBB.

See also: CostFunction, LSDBCost, BBCost

WaveletsExt.BestBasis.LSDB — Type

LSDB([; cost=DifferentialEntropyCost(), redundant=false])

Least Statistically Dependent Basis (LSDB). Set redundant=true when running LSDB with redundant wavelet transforms such as SWT or ACWT.

See also: BestBasisType, JBB, BB, SIBB

WaveletsExt.BestBasis.LSDBCost — Type

LSDBCost <: CostFunction

Cost function abstract type specifically for LSDB.

See also: CostFunction, JBBCost, BBCost

WaveletsExt.BestBasis.LogEnergyEntropyCost — Type

LogEnergyEntropyCost <: LSDBCost

Log energy entropy cost used for BB.

See also: CostFunction, BBCost, ShannonEntropyCost

WaveletsExt.BestBasis.LoglpCost — Type

LoglpCost <: JBBCost

$\log \ell^p$ information cost used for JBB. Typically, we set p=2 as in Wickerhauser's original algorithm.

See also: CostFunction, JBBCost, NormCost

WaveletsExt.BestBasis.NormCost — Type

NormCost <: JBBCost

$p$-norm information cost used for JBB.

See also: CostFunction, JBBCost, LoglpCost

WaveletsExt.BestBasis.SIBB — Type

SIBB([; cost=ShannonEntropyCost()])

Shift Invariant Best Basis (SIBB).

See also: BestBasisType, LSDB, JBB, BB

WaveletsExt.BestBasis.ShannonEntropyCost — Type

ShannonEntropyCost <: LSDBCost

Shannon entropy cost used for BB.

See also: CostFunction, BBCost, LogEnergyEntropyCost

Wavelets.Threshold.bestbasistree — Method

bestbasistree(X[, method])

Extension to the best basis tree function from Wavelets.jl. Given a set of decomposed signals, returns different types of best basis trees based on the methods specified. Available methods are the joint best basis (JBB), least statistically dependent basis (LSDB), individual regular best basis (BB), and shift-invariant best basis (SIBB).

Examples

bestbasistree(X, JBB())

bestbasistree(X, SIBB())

See also: bestbasiscoef

WaveletsExt.BestBasis.bestbasis_treeselection — Method

bestbasis_treeselection(costs, n[, type=:min])

Computes the best tree based on the given cost vector.

See also: bestbasistree, tree_costs

WaveletsExt.BestBasis.bestbasiscoef — Method

bestbasiscoef(X, tree)

bestbasiscoef(X, wt, tree)

Returns the expansion coefficients based on the given tree(s) and wavelet packet decomposition (WPD) expansion coefficients. If the WPD expansion coefficients were not provided, the expansion coefficients can be obtained by providing the signals and wavelet.

See also: bestbasistree, bestbasis_treeselection

WaveletsExt.BestBasis.tree_costs — Method

tree_costs(X, method)

Returns the cost of each node in a binary tree in order to find the best basis.

See also: bestbasistree, bestbasis_treeselection