Best Basis

WaveletsExt.BestBasis.bestbasis_treeselection

WaveletsExt.BestBasis.bestbasiscoef

WaveletsExt.BestBasis.tree_costs

Wavelets.Threshold.bestbasistree — Method

bestbasistree(X[, method])

Extension to the best basis tree function from Wavelets.jl. Given a set of decomposed signals, returns different types of best basis trees based on the methods specified. Available methods are the joint best basis (JBB()), least statistically dependent basis (LSDB()), and individual regular best basis (BB()).

WaveletsExt.BestBasis.bestbasis_treeselection — Method

bestbasis_treeselection(costs, n[, type=:min])

Computes the best tree based on the given cost vector.

WaveletsExt.BestBasis.bestbasiscoef — Method

bestbasiscoef(X, tree)
bestbasiscoef(X, wt, tree)

Returns the expansion coefficients based on the given tree(s) and wavelet packet decomposition (WPD) expansion coefficients. If the WPD expansion coefficients were not provided, the expansion coefficients can be obtained by providing the signals and wavelet.

WaveletsExt.BestBasis.tree_costs — Method

tree_costs(X, method)

Returns the cost of each node in a binary tree in order to find the best basis.