Denoising

WaveletsExt.Denoising.RelErrorShrink

WaveletsExt.Denoising.RelErrorShrink

WaveletsExt.Denoising.SureShrink

WaveletsExt.Denoising.SureShrink

WaveletsExt.Denoising.denoiseall

WaveletsExt.Denoising.findelbow

WaveletsExt.Denoising.orth2relerror

WaveletsExt.Denoising.relerrorthreshold

WaveletsExt.Denoising.surethreshold

Wavelets.Threshold.VisuShrink — Method

VisuShrink(n, th)

Extension to the VisuShrink struct constructor from Wavelets.jl.

WaveletsExt.Denoising.RelErrorShrink — Type

RelErroShrink(th, t) <: DNFT

Relative Error Shrink method used in their paper "Efficient Approximation and Denoising of Graph Signals using the Multiscale Basis Dictionary" for IEEE Transactions on Signal and Information Processing over Networks, Vol 0, No. 0,

WaveletsExt.Denoising.RelErrorShrink — Type

RelErrorShrink([th=HardTH()])

Struct constructor for Relative Error Shrink.

WaveletsExt.Denoising.SureShrink — Type

SureShrink(th, t) <: DNFT

Stein's Unbiased Risk Estimate (SURE) Shrink

WaveletsExt.Denoising.SureShrink — Type

SureShrink(x[, tree=nothing, th=SteinTH()])

Struct constructor for SureShrink based on the signal coefficients x.

See also: surethreshold

Wavelets.Threshold.denoise — Method

denoise(x, inputtype, wt[; L=maxtransformlevels(size(x,1)),
    tree=maketree(size(x,1), L, :dwt), dnt=VisuShrink(size(x,1)),
    estnoise=noisest, smooth=:regular])

Extension of the denoise function from Wavelets.jl. Denoise a signal of input type inputtype.

Arguments:

x::AbstractArray{<:Number}: input signals/coefficients.

inputtype::Symbol: input type of x. Current accepted types of inputs are

:sig: original signals; x should be a 2-D array with each column representing a signal.

:dwt: dwt-transformed signal coefficients; x should be a 1-D array with each column representing the coefficients of a signal.

:wpt: wpt-transformed signal coefficients; x should be a 1-D array with each column representing the coefficients of a signal.

:sdwt: sdwt-transformed signal coefficients; x should be a 2-D array with each column representing the coefficients of a node.

:swpd: swpd-transformed signal coefficients; x should be a 2-D array with each column representing the coefficients of a node.

:acwt: acwt-transformed signal coefficients from AutocorrelationShell.jl; x should be a 2-D array with each column representing the coefficients of a node.

:acwpt: acwpt-transformed signal coefficients from AutocorrelationShell.jl; x should be a 2-D array with each column representing the coefficients of a node.

wt::Union{DiscreteWavelet, Nothing}: the discrete wavelet to be used for decomposition (for input type :sig) and reconstruction. nothing can be supplied if no reconstruction is necessary.

L::Integer: the number of decomposition levels. Necessary for input types :sig, :dwt, and :sdwt. Default value is set to be maxtransformlevels(size(x,1)).

tree::BitVector: the decomposition tree of the signals. Necessary for input types :wpt and :swpd. Default value is set to be maketree(size(x,1), L, :dwt).

dnt::DNFT: denoise type. Default type is set to be VisuShrink(size(x,1)).

estnoise::Union{Function, Vector{<:Number}}: noise estimation. Input can be provided as a function to estimate noise in signal, or a vector of estimated noise. Default is set to be the noisest function.

smooth::Symbol: the smoothing method used. :regular smoothing thresholds all given coefficients, whereas :undersmooth smoothing does not threshold the lowest frequency subspace node of the wavelet transform. Default is set to be :regular.

See also: denoiseall, noisest, relerrorthreshold

Wavelets.Threshold.noisest — Method

noisest(x, redundant[, tree=nothing])

Extension to the noisest function from Wavelets.jl. Estimates the noise of a signal from its coefficients.

Examples

x = randn(128)
wt = wavelet(WT.haar)

# noise estimate for dwt transformation
y = dwt(x, wt)
noise = noisest(y, false)

# noise estimate for wpt transformation
tree = maketree(x, :full)
y = wpt(x, wt, tree)
noise = noisest(y, false, tree)

# noise estimate for sdwt transformation
y = sdwt(x, wt)
noise = noisest(y, true)

# noise estimate for swpd transformation
y = swpd(x, wt)
noise = noisest(y, true, tree)

See also: relerrorthreshold, VisuShrink, SureShrink

WaveletsExt.Denoising.denoiseall — Method

denoiseall(x, inputtype, wt[; L=maxtransformlevels(size(x,1)),
    tree=maketree(size(x,1), L, :dwt), dnt=VisuShrink(size(x,1)),
    estnoise=noisest, bestTH=nothing, smooth=:regular])

Denoise multiple signals of input type inputtype.

Arguments:

x::AbstractArray{<:Number}: input signals/coefficients.

inputtype::Symbol: input type of x. Current accepted types of inputs are

:sig: original signals; x should be a 2-D array with each column representing a signal.

:dwt: dwt-transformed signal coefficients; x should be a 2-D array with each column representing the coefficients of a signal.

:wpt: wpt-transformed signal coefficients; x should be a 2-D array with each column representing the coefficients of a signal.

:sdwt: sdwt-transformed signal coefficients; x should be a 3-D array with each 2-D slice representing the coefficients of a signal.

:swpd: swpd-transformed signal coefficients; x should be a 3-D array with each 2-D slice representing the coefficients of a signal.

:acwt: acwt-transformed signal coefficients from AutocorrelationShell.jl; x should be a 3-D array with each 2-D slice representing the coefficients of a signal.

:acwpt: acwpt-transformed signal coefficients from AutocorrelationShell.jl; x should be a 3-D array with each 2-D slice representing the coefficients of a signal.

wt::Union{DiscreteWavelet, Nothing}: the discrete wavelet to be used for decomposition (for input type :sig) and reconstruction. nothing can be supplied if no reconstruction is necessary.

L::Integer: the number of decomposition levels. Necessary for input types :sig, :dwt, and :sdwt. Default value is set to be maxtransformlevels(size(x,1)).

tree::BitVector: the decomposition tree of the signals. Necessary for input types :wpt and :swpd. Default value is set to be maketree(size(x,1), L, :dwt).

dnt::DNFT: denoise type. Default type is set to be VisuShrink(size(x,1)).

estnoise::Union{Function, Vector{<:Number}}: noise estimation. Input can be provided as a function to estimate noise in signal, or a vector of estimated noise. Default is set to be the noisest function.

bestTH::Union{Function, Nothing}: method to determine the best threshold value for a group of signals. If nothing is given, then each signal will be denoised by its respective best threshold value determined from the parameters dnt and estnoise; otherwise some function can be passed to determine the best threshold value from a vector of threshold values, eg: mean and median. Default is set to be nothing.

smooth::Symbol: the smoothing method used. :regular smoothing thresholds all given coefficients, whereas :undersmooth smoothing does not threshold the lowest frequency subspace node of the wavelet transform. Default is set to be :regular.

See alse: denoise, noisest, relerrorthreshold

WaveletsExt.Denoising.findelbow — Method

findelbow(x, y)

Given the x and y coordinates of a curve, return the elbow.

See also: RelErrorShrink, relerrorthreshold, orth2relerror

WaveletsExt.Denoising.orth2relerror — Method

orth2relerror(orth)

Given a vector 'orth' of orthonormal expansion coefficients, return a vector of relative approximation errors when retaining the 1,2,...,N largest coefficients in magnitude.

See also: RelErrorShrink, relerrorthreshold, findelbow

WaveletsExt.Denoising.relerrorthreshold — Method

relerrorthreshold(coef, redundant[, tree, elbows=2; makeplot=false])

Takes in a set of expansion coefficients, 'plot' the threshold vs relative error curve and select the best threshold value based on the elbow method. If one wants to see the resulting plot from this computation, simply set makeplot=true.

Examples

x = randn(128)
wt = wavelet(WT.haar)

# noise estimate for dwt transformation
y = dwt(x, wt)
noise = relerrorthreshold(y, false)

# noise estimate for wpt transformation
tree = maketree(x, :full)
y = wpt(x, wt, tree)
noise = relerrorthreshold(y, false, tree)

# noise estimate for sdwt transformation
y = sdwt(x, wt)
noise = relerrorthreshold(y, true)

# noise estimate for swpd transformation
y = swpd(x, wt)
noise = relerrorthreshold(y, true, tree)

See also: noisest, RelErrorShrink

WaveletsExt.Denoising.surethreshold — Method

surethreshold(coef, redundant[, tree=nothing])

Determination of the t value used for SureShrink.

See also: SureShrink