Local Discriminant Basis

WaveletsExt.LDB.LocalDiscriminantBasis

WaveletsExt.LDB.LocalDiscriminantBasis

WaveletsExt.LDB.change_nfeatures

WaveletsExt.LDB.discriminant_measure

WaveletsExt.LDB.discriminant_power

WaveletsExt.LDB.energy_map

WaveletsExt.LDB.fit!

WaveletsExt.LDB.fit_transform

WaveletsExt.LDB.inverse_transform

WaveletsExt.LDB.transform

WaveletsExt.LDB.LocalDiscriminantBasis — Type

LocalDiscriminantBasis

Class type for the Local Discriminant Basis (LDB), a feature selection algorithm developed by N. Saito and R. Coifman in "Local Discriminant Bases and Their Applications" in the Journal of Mathematical Imaging and Vision, Vol 5, 337-358 (1995). This struct contains the following field values:

Parameters and Attributes:

wt::DiscreteWavelet: a discrete wavelet for transform purposes

max_dec_level::Union{Integer, Nothing}: max level of wavelet packet decomposition to be computed.

dm::DiscriminantMeasure: the discriminant measure for the LDB algorithm. Supported measures are the AsymmetricRelativeEntropy(), LpEntropy(), SymmetricRelativeEntropy(), and HellingerDistance()

en::EnergyMap: the type of energy map used. Supported maps are TimeFrequency() and ProbabilityDensity().

dp::DiscriminantPower(): the measure of discriminant power among expansion coefficients. Supported measures are BasisDiscriminantMeasure(), FishersClassSeparability(), and RobustFishersClassSeparability().

top_k::Union{Integer, Nothing}: the top-k coefficients used in each node to determine the discriminant measure.

n_features::Union{Integer, Nothing}: the dimension of output after undergoing feature selection and transformation.

n::Union{Integer, Nothing}: length of signal

Γ::Union{AbstractArray{<:AbstractFloat}, Nothing}: computed energy map

DM::Union{AbstractArray{<:AbstractFloat}, Nothing}: computed discriminant measure

cost::Union{AbstractVector{<:AbstractFloat}, Nothing}: computed wavelet packet decomposition (WPD) tree cost based on the discriminant measure DM.

tree::Union{BitVector, Nothing}: computed best WPD tree based on the discriminant measure DM.

DP::Union{AbstractVector{<:AbstractFloat}, Nothing}: computed discriminant power

order::Union{AbstractVector{Integer}, Nothing}: ordering of DP by descending order.

WaveletsExt.LDB.LocalDiscriminantBasis — Method

LocalDiscriminantBasis(wt[; max_dec_level=nothing,
    dm=AsymmetricRelativeEntropy(), em=TimeFrequency(), 
    dp=BasisDiscriminantMeasure(), top_k=nothing,
    n_features=nothing])

Class constructor for LocalDiscriminantBasis.

Arguments:

wt::DiscreteWavelet: Wavelet used for decomposition of signals.

max_dec_level::Union{Integer, Nothing}: max level of wavelet packet decomposition to be computed. When max_dec_level=nothing, the maximum transform levels will be used. Default is set to be nothing.

dm::DiscriminantMeasure: the discriminant measure for the LDB algorithm. Supported measures are the AsymmetricRelativeEntropy(), LpEntropy(), SymmetricRelativeEntropy(), and HellingerDistance(). Default is set to be AsymmetricRelativeEntropy().

en::EnergyMap: the type of energy map used. Supported maps are TimeFrequency() and ProbabilityDensity(). Default is set to be TimeFrequency().

dp::DiscriminantPower=BasisDiscriminantMeasure(): the measure of discriminant power among expansion coefficients. Supported measures are BasisDiscriminantMeasure(), FishersClassSeparability(), and RobustFishersClassSeparability(). Default is set to be BasisDiscriminantMeasure().

top_k::Union{Integer, Nothing}: the top-k coefficients used in each node to determine the discriminant measure. When top_k=nothing, all coefficients are used to determine the discriminant measure. Default is set to be nothing.

n_features::Union{Integer, Nothing}: the dimension of output after undergoing feature selection and transformation. When n_features=nothing, all features will be returned as output. Default is set to be nothing.

WaveletsExt.LDB.change_nfeatures — Method

change_nfeatures(f, x, n_features)

Change the number of features from f.n_features to n_features.

Note that if the input n_features is larger than f.n_features, it results in the regeneration of signals based on the current f.n_features before reselecting the features. This will cause additional features to be less accurate and effective.

WaveletsExt.LDB.discriminant_measure — Method

discriminant_measure(Γ, dm)

Returns the discriminant measure calculated from the energy maps.

WaveletsExt.LDB.discriminant_power — Method

discriminant_power(coefs, tree, dp)

Returns the discriminant power of each leaf from the local discriminant basis (LDB) tree.

WaveletsExt.LDB.energy_map — Method

energy_map(coef, method)

Returns the Time Frequency Energy map or the Probability Density Energy map depending on the input method (TimeFrequency() or ProbabilityDensity()).

WaveletsExt.LDB.fit! — Method

fit!(f, X, y)

Fits the Local Discriminant Basis feature selection algorithm f onto the signals X (or the decomposed signals Xw) with labels y.

WaveletsExt.LDB.fit_transform — Method

fit_transform(f, X, y)

Fit and transform the signals X with labels y based on the LDB class f.

WaveletsExt.LDB.inverse_transform — Method

inverse_transform(f, x)

Compute the inverse transform on the feature matrix x to form the original signal based on the LDB class f.

WaveletsExt.LDB.transform — Method

transform(f, X)

Extract the LDB features on signals X.