Local Discriminant Basis

WaveletsExt.LDB.AsymmetricRelativeEntropy

WaveletsExt.LDB.BasisDiscriminantMeasure

WaveletsExt.LDB.DiscriminantMeasure

WaveletsExt.LDB.DiscriminantPower

WaveletsExt.LDB.EarthMoverDistance

WaveletsExt.LDB.EnergyMap

WaveletsExt.LDB.FishersClassSeparability

WaveletsExt.LDB.HellingerDistance

WaveletsExt.LDB.LocalDiscriminantBasis

WaveletsExt.LDB.LocalDiscriminantBasis

WaveletsExt.LDB.LpDistance

WaveletsExt.LDB.ProbabilityDensity

WaveletsExt.LDB.ProbabilityDensityDM

WaveletsExt.LDB.RobustFishersClassSeparability

WaveletsExt.LDB.Signatures

WaveletsExt.LDB.SignaturesDM

WaveletsExt.LDB.SymmetricRelativeEntropy

WaveletsExt.LDB.TimeFrequency

WaveletsExt.LDB.change_nfeatures

WaveletsExt.LDB.discriminant_measure

WaveletsExt.LDB.discriminant_power

WaveletsExt.LDB.energy_map

WaveletsExt.LDB.fit!

WaveletsExt.LDB.fit_transform

WaveletsExt.LDB.inverse_transform

WaveletsExt.LDB.transform

Public API

WaveletsExt.LDB.AsymmetricRelativeEntropy — Type

AsymmetricRelativeEntropy <: ProbabilityDensityDM

Asymmetric Relative Entropy discriminant measure for the Probability Density and Time Frequency based energy maps. This measure is also known as cross entropy and Kullback-Leibler divergence.

Equation: $D(p,q) = \sum p(x) \log \frac{p(x)}{q(x)}$

WaveletsExt.LDB.BasisDiscriminantMeasure — Type

BasisDiscriminantMeasure <: DiscriminantPower

This is the discriminant measure of a single basis function computed in a previous step to construct the energy maps.

WaveletsExt.LDB.DiscriminantMeasure — Type

Discriminant measure for Local Discriminant Basis. Current available subtypes are:

ProbabilityDensityDM

SignaturesDM

WaveletsExt.LDB.DiscriminantPower — Type

Discriminant Power measure for the Local Discriminant Basis. Current available measures are

BasisDiscriminantMeasure

FishersClassSeparability

RobustFishersClassSeparability

WaveletsExt.LDB.EarthMoverDistance — Type

EarthMoverDistance <: SignaturesDM

Earth Mover Distance discriminant measure for the Signatures energy map.

Equation: $E(P,Q) = \frac{\sum_{k=1}^{m+n+1} |\hat p_k - \hat q_k| (r_{k+1} - r_k)}{w_\Sigma}$

where $r_1, r_2, \ldots, r_{m+n}$ is the sorted list of $p_1, \ldots, p_m, q_1, \ldots, q_n$ and $\hat p_k = \sum_{p_i \leq r_k} w_{p_i}$, $\hat q_k = \sum_{q)i \leq r_k} w_{q_i}$.

WaveletsExt.LDB.EnergyMap — Type

Energy map for Local Discriminant Basis. Current available types are:

TimeFrequency

ProbabilityDensity

WaveletsExt.LDB.FishersClassSeparability — Type

FishersClassSeparability <: DiscriminantPower

The Fisher's class separability of the expansion coefficients in the basis function.

Equation: $\frac{\sum_{c=1}^C \pi_c({\rm mean}_i(\alpha_{\lambda,i}^{(c)}) - {\rm mean}_c({\rm mean}_i(\alpha_{\lambda,i}^{(c)})))^2}{\sum_{c=1}^C \pi_c {\rm var}_i(\alpha_{\lambda,i}^{(c)})}$

WaveletsExt.LDB.HellingerDistance — Type

HellingerDistance <: ProbabilityDensityDM

Hellinger Distance discriminant measure for the Probability Density energy map.

Equation: $H(p,q) = \sum_{i=1}^n (\sqrt{p_i} - \sqrt{q_i})^2$

WaveletsExt.LDB.LocalDiscriminantBasis — Type

LocalDiscriminantBasis

Class type for the Local Discriminant Basis (LDB), a feature selection algorithm developed by N. Saito and R. Coifman in "Local Discriminant Bases and Their Applications" in the Journal of Mathematical Imaging and Vision, Vol 5, 337-358 (1995). This struct contains the following field values:

Parameters and Attributes:

wt::DiscreteWavelet: a discrete wavelet for transform purposes

max_dec_level::Union{Integer, Nothing}: max level of wavelet packet decomposition to be computed.

dm::DiscriminantMeasure: the discriminant measure for the LDB algorithm. Supported measures are the AsymmetricRelativeEntropy(), LpDistance(), SymmetricRelativeEntropy(), and HellingerDistance()

en::EnergyMap: the type of energy map used. Supported maps are TimeFrequency(), ProbabilityDensity(), and Signatures().

dp::DiscriminantPower(): the measure of discriminant power among expansion coefficients. Supported measures are BasisDiscriminantMeasure(), FishersClassSeparability(), and RobustFishersClassSeparability().

top_k::Union{Integer, Nothing}: the top-k coefficients used in each node to determine the discriminant measure.

n_features::Union{Integer, Nothing}: the dimension of output after undergoing feature selection and transformation.

n::Union{Integer, Nothing}: length of signal

`Γ::Union{AbstractArray{<:AbstractFloat},

AbstractArray{NamedTuple{(:coef, :weight), Tuple{S1, S2}}} where {S1<:Array{T} where T<:AbstractFloat, S2<:Union{AbstractFloat, Array{<:AbstractFloat}}}, Nothing}`: computed energy map

DM::Union{AbstractArray{<:AbstractFloat}, Nothing}: computed discriminant measure

cost::Union{AbstractVector{<:AbstractFloat}, Nothing}: computed wavelet packet decomposition (WPD) tree cost based on the discriminant measure DM.

tree::Union{BitVector, Nothing}: computed best WPD tree based on the discriminant measure DM.

DP::Union{AbstractVector{<:AbstractFloat}, Nothing}: computed discriminant power

order::Union{AbstractVector{Integer}, Nothing}: ordering of DP by descending order.

WaveletsExt.LDB.LocalDiscriminantBasis — Method

LocalDiscriminantBasis([; 
    wt=wavelet(WT.haar),
    max_dec_level=nothing,
    dm=AsymmetricRelativeEntropy(), em=TimeFrequency(), 
    dp=BasisDiscriminantMeasure(), top_k=nothing,
    n_features=nothing]
)

Class constructor for LocalDiscriminantBasis.

Arguments:

wt::DiscreteWavelet: Wavelet used for decomposition of signals. Default is set to be wavelet(WT.haar).

max_dec_level::Union{Integer, Nothing}: max level of wavelet packet decomposition to be computed. When max_dec_level=nothing, the maximum transform levels will be used. Default is set to be nothing.

dm::DiscriminantMeasure: the discriminant measure for the LDB algorithm. Supported measures are the AsymmetricRelativeEntropy(), LpDistance(), SymmetricRelativeEntropy(), and HellingerDistance(). Default is set to be AsymmetricRelativeEntropy().

en::EnergyMap: the type of energy map used. Supported maps are TimeFrequency() and ProbabilityDensity(). Default is set to be TimeFrequency().

dp::DiscriminantPower=BasisDiscriminantMeasure(): the measure of discriminant power among expansion coefficients. Supported measures are BasisDiscriminantMeasure(), FishersClassSeparability(), and RobustFishersClassSeparability(). Default is set to be BasisDiscriminantMeasure().

top_k::Union{Integer, Nothing}: the top-k coefficients used in each node to determine the discriminant measure. When top_k=nothing, all coefficients are used to determine the discriminant measure. Default is set to be nothing.

n_features::Union{Integer, Nothing}: the dimension of output after undergoing feature selection and transformation. When n_features=nothing, all features will be returned as output. Default is set to be nothing.

WaveletsExt.LDB.LpDistance — Type

LpDistance <: ProbabilityDensityDM

$\ell^p$ Distance discriminant measure for the Probability Density and Time Frequency based energy maps. The default $p$ value is set to 2.

Equation: $W(q,r) = ||q-r||_p^p = \sum_{i=1}^n (q_i - r_i)^p$

WaveletsExt.LDB.ProbabilityDensity — Type

ProbabilityDensity <: EnergyMap

An energy map based on probability density, a measure based on the differences among the pdfs of $Z_i$. Since we do not know the true density functions of the coefficients, the PDFs are estimated using the Average Shifted Histogram (ASH).

See also: EnergyMap, TimeFrequency, Signatures

WaveletsExt.LDB.ProbabilityDensityDM — Type

Discriminant measure for Probability Density and Time Frequency based energy maps. Current available measures are:

AsymmetricRelativeEntropy

SymmetricRelativeEntropy

LpDistance

HellingerDistance

WaveletsExt.LDB.RobustFishersClassSeparability — Type

RobustFishersClassSeparability <: DiscriminantPower

The robust version of Fisher's class separability of the expansion coefficients in the basis function.

Equation: $\frac{\sum_{c=1}^C \pi_c({\rm med}_i(\alpha_{\lambda,i}^{(c)}) - {\rm med}_c({\rm med}_i(\alpha_{\lambda,i}^{(c)})))^2}{\sum_{c=1}^C \pi_c {\rm mad}_i(\alpha_{\lambda,i}^{(c)})}$

WaveletsExt.LDB.Signatures — Type

Signatures <: EnergyMap

An energy map based on signatures, a measure that uses the Earth Mover's Distance (EMD) to compute the discriminating power of a coordinate. Signatures provide us with a fully data-driven representation, which can be efficiently used with EMD. This representation is more efficient than a histogram and is able to represent complex data structure with fewer samples.

Here, a signature for the coefficients in the $j$-th level, $k$-th node, $l$-th index of class $c$ is defined as

$s_{j,k,l}^{(c)} = \{(\alpha_{i;j,k,l}^{(c)}, w_{i;j,k,l}^{(c)})\}_{i=1}^{N_c}$

where $\alpha_{i;j,k,l}^{(c)}$ and $w_{i;j,k,l}^{(c)}$ are the expansion coefficients and weights at location $(j,k,l)$ for signal $i$ of class $c$ respectively. Currently, the two valid types of weights are :equal and :pdf.

Argumemts

weight::Symbol: Type of weight to be used to compute $w_{i;j,k,l}^{(c)}$. Available methods are :equal and pdf. Default is set to :equal.

See also: EnergyMap, TimeFrequency, ProbabilityDensity

WaveletsExt.LDB.SignaturesDM — Type

Discriminant measure for Signatures based energy maps. Current available measures are:

EarthMoverDistance

WaveletsExt.LDB.SymmetricRelativeEntropy — Type

SymmetricRelativeEntropy <: ProbabilityDensityDM

Symmetric Relative Entropy discriminant measure for the Probability Density and Time Frequency energy maps. Similar idea to the Asymmetric Relative Entropy, but this aims to make the measure more symmetric.

Equation: Denote the Asymmetric Relative Entropy as $D_A(p,q)$, then

$D(p,q) = D_A(p,q) + D_A(q,p) = \sum p(x) \log \frac{p(x)}{q(x)} + q(x) \log \frac{q(x)}{p(x)}$

WaveletsExt.LDB.TimeFrequency — Type

TimeFrequency <: EnergyMap

An energy map based on time frequencies, a measure based on the differences of derived quantities from projection $Z_i$, such as mean class energies or cumulants.

See also: EnergyMap, ProbabilityDensity, Signatures

WaveletsExt.LDB.change_nfeatures — Method

change_nfeatures(f, x, n_features)

Change the number of features from f.n_features to n_features.

Note that if the input n_features is larger than f.n_features, it results in the regeneration of signals based on the current f.n_features before reselecting the features. This will cause additional features to be less accurate and effective.

See also: LocalDiscriminantBasis, fit!, fit_transform, transform, inverse_transform

WaveletsExt.LDB.discriminant_measure — Method

discriminant_measure(Γ, dm)

Returns the discriminant measure of each node calculated from the energy maps.

WaveletsExt.LDB.discriminant_power — Method

discriminant_power(D, tree, dp)

Returns the discriminant power of each leaf from the local discriminant basis (LDB) tree.

WaveletsExt.LDB.energy_map — Method

energy_map(Xw, y, method)

Returns the Time Frequency Energy map or the Probability Density Energy map depending on the input method (TimeFrequency() or ProbabilityDensity()).

See also: EnergyMap. TimeFrequency, ProbabilityDensity

WaveletsExt.LDB.fit! — Method

fit!(f, X, y)

Fits the Local Discriminant Basis feature selection algorithm f onto the signals X (or the decomposed signals Xw) with labels y.

See also: LocalDiscriminantBasis, fit_transform, transform, inverse_transform, change_nfeatures

WaveletsExt.LDB.fit_transform — Method

fit_transform(f, X, y)

Fit and transform the signals X with labels y based on the LDB class f.

See also: LocalDiscriminantBasis, fit!, transform, inverse_transform, change_nfeatures

WaveletsExt.LDB.inverse_transform — Method

inverse_transform(f, x)

Compute the inverse transform on the feature matrix x to form the original signal based on the LDB class f.

See also: LocalDiscriminantBasis, fit!, fit_transform, transform, change_nfeatures

WaveletsExt.LDB.transform — Method

transform(f, X)

Extract the LDB features on signals X.

See also: LocalDiscriminantBasis, fit!, fit_transform, inverse_transform, change_nfeatures

Private API