Stationary Wavelet Transform

WaveletsExt.SWT.isdwt

WaveletsExt.SWT.isdwt_step

WaveletsExt.SWT.isdwt_step!

WaveletsExt.SWT.iswpt

WaveletsExt.SWT.sdwt

WaveletsExt.SWT.sdwt_step

WaveletsExt.SWT.sdwt_step!

WaveletsExt.SWT.swpd

WaveletsExt.SWT.swpt

WaveletsExt.SWT.isdwt — Function

isdwt(xw, wt[, ε])

Performs the inverse stationary discrete wavelet transform (ISDWT) on the sdwt transform coefficients with respect to the Boolean Vector ε which represents the shifts to be used. If ε is not provided, the average-basis ISDWT will be computed instead.

Examples

# decompose signal
xw = sdwt(x, wt, 5)

# ε-based reconstruction
y = isdwt(xw, wt, BitVector([0,1,0,0,0]))

# average-based reconstruction
y = isdwt(xw, wt)

See also: iswpt, sdwt

WaveletsExt.SWT.isdwt_step — Function

isdwt_step(v1, w1, j, s0, s1, g, h)

Perform one level of the inverse stationary discrete wavelet transform (ISDWT) on the vector v1 and w1, which is the j-th level scaling coefficients (Note the 0th level scaling coefficients is the raw signal). The vectors h and g are the detail and scaling filters.

Returns a vector v0 of the scaling and detail coefficients at level j-1.

See also: isdwt_step!

WaveletsExt.SWT.isdwt_step! — Function

isdwt_step!(v0, v1, w1, j, s0, s1, g, h)

Same as isdwt_step but without array allocation.

See also: isdwt_step

WaveletsExt.SWT.iswpt — Function

iswpt(xw, wt, ε[, L=maxtransformlevels(size(xw,1))])

iswpt(xw, wt, ε, tree)

iswpt(xw, wt[, L=maxtransformlevels(size(xw,1))])

iswpt(xw, wt, tree)

Performs the inverse stationary wavelet packet transform (ISWPT) on the swpd transform coefficients with respect to a given Boolean Vector that represents a binary tree and the BitVector ε which represents the shifts to be used. If ε is not provided, the average-basis ISWPT will be computed instead.

Examples

# decompose signal
xw = swpd(x, wt, 5)

# best basis tree
bt = bestbasistree(xw, BB(redundant=true))

# ε-based reconstruction
y = iswpt(xw, wt, BitVector([0,0,0,1,0]), 5)
y = iswpt(xw, wt, BitVector([0,0,0,1,0]), bt)

# average-based reconstruction
y = iswpt(xw, wt, 5)
y = iswpt(xw, wt, bt)

See also: isdwt, swpd

WaveletsExt.SWT.sdwt — Function

sdwt(x, wt[, L=maxtransformlevels(x)])

Perform a stationary discrete wavelet transform (SDWT) of the array x. The wavelet type wt determines the transform type and the wavelet class, see wavelet.

The number of transform levels L can be 1 ≤ L ≤ maxtransformlevels(x). Default value is set to maxtransformlevels(x).

Returns the n × (L+1) matrix (where n is the length of x) with the detail coefficients for level j in column (L-j+2). The scaling coefficients are in the 1st column.

Examples

xw = sdwt(x, wt, 5)

See also: swpd, swpt, isdwt

WaveletsExt.SWT.sdwt_step — Function

sdwt_step(v, j, h, g)

Perform one level of the stationary discrete wavelet transform (SDWT) on the vector v, which is the j-th level scaling coefficients (Note the 0th level scaling coefficients is the raw signal). The vectors h and g are the detail and scaling filters.

Returns a tuple (v, w) of the scaling and detail coefficients at level j+1.

See also: sdwt_step!

WaveletsExt.SWT.sdwt_step! — Function

sdwt_step!(v1, w1, v, j, h, g)

Same as sdwt_step but without array allocation.

See also: sdwt_step

WaveletsExt.SWT.swpd — Function

swpd(x, wt[, L=maxtransformlevels(x)])

Perform a stationary wavelet packet decomposition (SPWD) of the array x. The wavelet type wt determines the transform type and the wavelet class, see wavelet.

The number of transform levels L can be 1 ≤ L ≤ maxtransformlevels(x). Default value is set to maxtransformlevels(x).

Returns the n × (2⁽ᴸ⁺¹⁾-1) matrix (where n is the length of x) with each column representing the nodes in the binary tree.

Examples

xw = swpd(x, wt, 5)

See also: swpt, sdwt, iswpt

WaveletsExt.SWT.swpt — Function

swpt(x, wt[, L=maxtransformlevels(x)])

swpt(x, wt, tree)

swpt(x, h, g, tree)

swpt(x, h, g, tree, i)

Performs the stationary wavelet packet transform (SWPT) of the vector x of length N = 2ᴸ. The wavelet type wt determines the transform type and the wavelet class, see wavelet.

The number of transform levels L can be 1 ≤ L ≤ maxtransformlevels(x). Default value is set to maxtransformlevels(x).

Returns the expansion coefficients of the SWPT of the size N × k. Each column represents a leaf node from tree. Number of returned columns can vary between 1 ≤ k ≤ N depending on the input tree.

Note: If one wants to compute the stationary wavelet packet transform on a signal, yet hopes to reconstruct the original signal later on, please use the swpd function instead.

See also: sdwt, swpd