Utils

WaveletsExt.Utils.ClassData

WaveletsExt.Utils.coarsestscalingrange

WaveletsExt.Utils.duplicatesignals

WaveletsExt.Utils.finestdetailrange

WaveletsExt.Utils.generateclassdata

WaveletsExt.Utils.generatesignals

WaveletsExt.Utils.getcolrange

WaveletsExt.Utils.getleaf

WaveletsExt.Utils.getquadtreelevel

WaveletsExt.Utils.getrowrange

WaveletsExt.Utils.left

WaveletsExt.Utils.makequadtree

WaveletsExt.Utils.nodelength

WaveletsExt.Utils.psnr

WaveletsExt.Utils.relativenorm

WaveletsExt.Utils.right

WaveletsExt.Utils.snr

WaveletsExt.Utils.ssim

WaveletsExt.Utils.ClassData — Type

ClassData(type, s₁, s₂, s₃)

Based on the input type, generates 3 classes of signals with sample sizes s₁, s₂, and s₃ respectively. Accepted input types are:

:tri: Triangular signals of length 32

:cbf: Cylinder-Bell-Funnel signals of length 128

Based on N. Saito and R. Coifman in "Local Discriminant Basis and their Applications" in the Journal of Mathematical Imaging and Vision, Vol. 5, 337-358 (1995).

See also: generateclassdata

Wavelets.Util.maxtransformlevels — Method

maxtransformlevels(n)

maxtransformlevels(x[, dims])

Extension function from Wavelets.jl. Finds the max number of transform levels for an array x. If dims is provided for a multidimensional array, then it finds the max transform level for that corresponding dimension.

Arguments

x::AbstractArray: Input array.

dims::Integer: Dimension used for wavelet transform.

n::Integer: Length of input vector.

Returns

::Integer: Max number of transform levels.

Examples

using Wavelets, WaveletsExt

# Define random signal
x = randn(64, 128)

# Max transform levels and their corresponding return values
maxtransformlevels(128)     # 7
maxtransformlevels(x)       # 6
maxtransformlevels(x, 2)    # 7

WaveletsExt.Utils.coarsestscalingrange — Method

coarsestscalingrange(x, tree[, redundant=false])

coarsestscalingrange(n, tree[, redundant=false])

Given a binary tree, returns the index range of the coarsest scaling coefficients.

WaveletsExt.Utils.duplicatesignals — Method

duplicatesignals(x, N, k[, noise=false, t=1])

Given a signal x, returns N shifted versions of the signal, each with shifts of multiples of k.

Setting noise = true allows randomly generated Gaussian noises of μ = 0, σ² = t to be added to the circularly shifted signals.

WaveletsExt.Utils.finestdetailrange — Method

finestdetailrange(x, tree[, redundant=false])

finestdetailrange(n, tree[, redundant=false])

Given a binary tree, returns the index range of the coarsest scaling coefficients.

WaveletsExt.Utils.generateclassdata — Function

generateclassdata(c[, shuffle=false])

Generates 3 classes of data given a ClassData struct as an input. Returns a matrix containing the 3 classes of signals and a vector containing their corresponding labels.

Based on N. Saito and R. Coifman in "Local Discriminant Basis and their Applications" in the Journal of Mathematical Imaging and Vision, Vol. 5, 337-358 (1995).

See also: ClassData

WaveletsExt.Utils.generatesignals — Method

generatesignals(fn, L)

Generates a signal of length 2ᴸ given the function symbol fn. Current accepted inputs below are based on D. Donoho and I. Johnstone in "Adapting to Unknown Smoothness via Wavelet Shrinkage" Preprint Stanford, January 93, p 27-28.

:blocks

:bumps

:heavysine

:doppler

:quadchirp

:mishmash

The code for this function is adapted and translated based on MATLAB's Wavelet Toolbox's wnoise function.

Examples

generatesignals(:bumps, 8)

WaveletsExt.Utils.getcolrange — Method

getcolrange(n, idx)

Get the column range from a matrix with n columns that corresponds to idx from a quadtree.

Arguments

n::Integer: Number of columns in matrix.

idx::T where T<:Integer: Index from a quadtree corresponding to the matrix.

Returns

::UnitRange{Int64}: Column range in matrix that corresponds to idx.

Examples

using WaveletsExt

x = randn(8,8)
tree = makequadtree(x, 3, :full)
getcolrange(8,3)            # 5:8

See also: makequadtree, getrowrange

WaveletsExt.Utils.getleaf — Method

getleaf(tree)

Returns the leaf nodes of a tree.

WaveletsExt.Utils.getquadtreelevel — Method

getquadtreelevel(idx)

Get level of idx in the quadtree.

Arguments

idx::T where T<:Integer: Index of a quadtree.

Returns

::T: Level of idx.

Examples

using WaveletsExt

getquadtreelevel(1)     # 0
getquadtreelevel(3)     # 1

See also: makequadtree

WaveletsExt.Utils.getrowrange — Method

getrowrange(n, idx)

Get the row range from a matrix with n rows that corresponds to idx from a quadtree.

Arguments

n::Integer: Number of rows in matrix.

idx::T where T<:Integer: Index from a quadtree corresponding to the matrix.

Returns

::UnitRange{Int64}: Row range in matrix that corresponds to idx.

Examples

using WaveletsExt

x = randn(8,8)
tree = makequadtree(x, 3, :full)
getrowrange(8,3)            # 1:4

See also: makequadtree, getcolrange

WaveletsExt.Utils.left — Method

left(i)

Given the node index i, returns the index of its left node.

See also: right

WaveletsExt.Utils.makequadtree — Method

makequadtree(x, L[, s])

Build quadtree for 2D wavelet transform. Indexing of the quadtree are as follows:

Level 0                 Level 1                 Level 2             ...
_________________       _________________       _________________
|               |       |   2   |   3   |       |_6_|_7_|10_|11_|
|       1       |       |_______|_______|       |_8_|_9_|12_|13_|   ...
|               |       |   4   |   5   |       |14_|15_|18_|19_|
|_______________|       |_______|_______|       |16_|17_|20_|21_|

Arguments

x::AbstractArray{T,2} where T<:Number: Input array.

L::Integer: Number of decomposition levels.

s::Symbol: (Default: :full) Type of quadtree. Available types are :full and :dwt.

Returns

::BitVector: Quadtree representation.

Examples

using WaveletsExt

x = randn(16,16)
makequadtree(x, 3)

WaveletsExt.Utils.nodelength — Method

nodelength(N, L)

Returns the node length at level L of a signal of length N. Level L == 0 corresponds to the original input signal.

WaveletsExt.Utils.psnr — Method

psnr(x, x₀)

Returns the peak signal to noise ratio (PSNR) between original signal x₀ and noisy signal x.

See also: relativenorm, snr, ssim

WaveletsExt.Utils.relativenorm — Method

relativenorm(x, x₀[, p=2]) where T<:Number

Returns the relative norm of base p between original signal x₀ and noisy signal x.

See also: psnr, snr, ssim

WaveletsExt.Utils.right — Method

right(i)

Given the node index i, returns the index of its right node.

See also: left

WaveletsExt.Utils.snr — Method

snr(x, x₀)

Returns the signal to noise ratio (SNR) between original signal x₀ and noisy signal x.

See also: relativenorm, psnr, ssim

WaveletsExt.Utils.ssim — Method

ssim(x, x₀)

Wrapper for assess_ssim function from ImageQualityIndex.jl.

Returns the Structural Similarity Index Measure (SSIM) between the original signal/image x₀ and noisy signal/image x.

See also: relativenorm, psnr, snr