Utils

WaveletsExt.Utils.coarsestscalingrange

WaveletsExt.Utils.finestdetailrange

WaveletsExt.Utils.generatesignals

WaveletsExt.Utils.getleaf

WaveletsExt.Utils.left

WaveletsExt.Utils.nodelength

WaveletsExt.Utils.psnr

WaveletsExt.Utils.relativenorm

WaveletsExt.Utils.right

WaveletsExt.Utils.snr

WaveletsExt.Utils.ssim

WaveletsExt.Utils.coarsestscalingrange — Method

coarsestscalingrange(x, tree[, redundant=false])

coarsestscalingrange(n, tree[, redundant=false])

Given a binary tree, returns the index range of the coarsest scaling coefficients.

WaveletsExt.Utils.finestdetailrange — Method

finestdetailrange(x, tree[, redundant=false])

finestdetailrange(n, tree[, redundant=false])

Given a binary tree, returns the index range of the coarsest scaling coefficients.

WaveletsExt.Utils.generatesignals — Method

generatesignals(x, N, k[, noise=false, t=1])

Given a signal x, returns N shifted versions of the signal, each with shifts of multiples of k.

Setting noise = true allows randomly generated Gaussian noises of μ = 0, σ² = t to be added to the circularly shifted signals.

WaveletsExt.Utils.getleaf — Method

getleaf(tree)

Returns the leaf nodes of a tree.

WaveletsExt.Utils.left — Method

left(i)

Given the node index i, returns the index of its left node.

WaveletsExt.Utils.nodelength — Method

nodelength(N, L)

Returns the node length at level L of a signal of length N. Level L == 0 corresponds to the original input signal.

WaveletsExt.Utils.psnr — Method

psnr(x, x₀)

Returns the peak signal to noise ratio (PSNR) between original signal x₀ and noisy signal x.

WaveletsExt.Utils.relativenorm — Method

relativenorm(x, x₀[, p=2]) where T<:Number

Returns the relative norm of base p between original signal x₀ and noisy signal x.

WaveletsExt.Utils.right — Method

right(i)

Given the node index i, returns the index of its right node.

WaveletsExt.Utils.snr — Method

snr(x, x₀)

Returns the signal to noise ratio (SNR) between original signal x₀ and noisy signal x.

WaveletsExt.Utils.ssim — Method

ssim(x, x₀)

Wrapper for assess_ssim function from ImageQualityIndex.jl.

Returns the Structural Similarity Index Measure (SSIM) between the original signal/image x₀ and noisy signal/image x.